% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_noise_tensor}
\alias{add_noise_tensor}
\title{Add noise to tensor}
\usage{
add_noise_tensor(x, noise_type, ...)
}
\arguments{
\item{noise_type}{"normal" or "uniform".}

\item{...}{additional arguments for rnorm or runif call.}
}
\description{
Add noise to tensor
}
