% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{auc_roc_metric}
\alias{auc_roc_metric}
\title{Computes mean and median auc scores for model with 1 or multiple binary classifiers in last layer.}
\usage{
auc_roc_metric(
  file_path,
  model = NULL,
  batch_size = 100,
  numberOfBatches = 10,
  return_auc_vector = FALSE,
  evaluate_all_files = FALSE,
  return_auprc_vector = FALSE,
  seed = 1234
)
}
\arguments{
\item{file_path}{Path to folder containing rds files.}

\item{model}{A keras model.}

\item{numberOfBatches}{How many batches to evaluate.}

\item{return_auc_vector}{Whether to return AUC vector containing scores for each class.}

\item{evaluate_all_files}{Whether to evaluate all samples in \code{file_path} excactly once.}

\item{return_auprc_vector}{Whether to return AUPRC vector containing scores for each class.}

\item{seed}{Sets seed for set.seed function, for reproducible results when using \code{randomFiles} or \code{shuffleFastaEntries}}
}
\description{
Computes mean and median auc scores for model with 1 or multiple binary classifiers in last layer.
}
