% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{balanced_acc_wrapper}
\alias{balanced_acc_wrapper}
\title{Balanced accuracy metric}
\usage{
balanced_acc_wrapper(num_targets, cm_dir)
}
\arguments{
\item{num_targets}{Number of targets.}

\item{cm_dir}{Directory of confusion matrix used to compute balanced accuracy.}
}
\description{
Compute balanced accuracy as additional score. Useful for imbalanced data.
}
