% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{calculate_steps_per_epoch}
\alias{calculate_steps_per_epoch}
\title{do one full preprocessing iteration to figure out what the observed
steps_per_epoch value is}
\usage{
calculate_steps_per_epoch(dir, batch_size = 12)
}
\description{
do one full preprocessing iteration to figure out what the observed
steps_per_epoch value is
}
