% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateStepsPerEpoch}
\alias{calculateStepsPerEpoch}
\title{Calculate the steps per epoch}
\usage{
calculateStepsPerEpoch(
  dir,
  batch.size = 256,
  maxlen = 250,
  format = "fasta",
  verbose = F
)
}
\arguments{
\item{dir}{Input directory where .fasta files are located}

\item{batch.size}{Number of samples}

\item{maxlen}{Length of the semi-redundant sequences}

\item{format}{File format}

\item{verbose}{TRUE/FALSE}
}
\description{
Do one full preprocessing iteration to the FASTA file to figure out what the observed
steps_per_epoch value is.
}
