% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{check_layer_name}
\alias{check_layer_name}
\title{Check if layer is in model}
\usage{
check_layer_name(model, layer_name)
}
\value{
Error message if model does not contain layer of certain name.
}
\description{
Check if layer is in model
}
\keyword{internal}
