% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{checkpoint_cb}
\alias{checkpoint_cb}
\title{Checkpoint callback}
\usage{
checkpoint_cb(
  filepath_checkpoints,
  save_weights_only,
  save_best_only,
  save_freq,
  monitor = "val_loss"
)
}
\arguments{
\item{save_weights_only}{Whether to save weights only.}

\item{save_best_only}{Only save model that improved on some score. Not applied if argument is \code{NULL}. Otherwise must be
list with argument \code{monitor} or \code{save_freq} (can only use one option). \code{moniter} specifies what metric to use.
\code{save_freq}, integer specifying how often to store a checkpoint (in epochs).}
}
\description{
Checkpoint callback
}
\keyword{internal}
