% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{cm_perc}
\alias{cm_perc}
\title{Transform confusion matrix with total numbers to matrix with percentages.}
\usage{
cm_perc(cm, round_dig = 2)
}
\description{
Transform confusion matrix with total numbers to matrix with percentages.
}
\keyword{internal}
