% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compute_gradients}
\alias{compute_gradients}
\title{Compute gradients}
\usage{
compute_gradients(
  input_tensor,
  target_class_idx,
  model,
  input_idx = NULL,
  pred_stepwise = FALSE
)
}
\arguments{
\item{target_class_idx}{Index of class to compute gradient for.}

\item{model}{Model to compute gradient for.}

\item{input_idx}{Input layer to monitor for > 1 input.}

\item{pred_stepwise}{Whether to do predictions with batch_size 1 rather than all at once. Can be used if
input is too big to handle at once.}
}
\description{
Compute gradients
}
