% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{count_nuc}
\alias{count_nuc}
\title{Count nucleotides per class}
\usage{
count_nuc(
  path,
  vocabulary_label = NULL,
  format = "fasta",
  file_proportion = 1,
  train_type = "label_folder",
  csv_path = NULL
)
}
\arguments{
\item{vocabulary_label}{Character vector of possible targets. Targets outside \code{vocabulary_label} will get discarded.}

\item{format}{File format, either \code{"fasta"} or \code{"fastq"}.}

\item{file_proportion}{Proportion of files to randomly sample for estimating class distributions.}

\item{train_type}{Either "label_folder", "label_header" or "label_csv".}
}
\description{
Count nucleotides per class
}
\keyword{internal}
