% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dummy_data.R
\name{create_dummy_data}
\alias{create_dummy_data}
\title{create random sequences and write to fasta file}
\usage{
create_dummy_data(file_path, num_files, seq_length, num_seq, vocabulary)
}
\description{
create random sequences and write to fasta file
}
