% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model_wavenet}
\alias{create_model_wavenet}
\title{create wavenet model}
\usage{
create_model_wavenet(
  filters = 16,
  kernel_size = 2,
  residual_blocks,
  maxlen,
  input_tensor = NULL,
  initial_kernel_size = 32,
  initial_filters = 32,
  output_channels = 4,
  output_activation = "softmax",
  solver = "adam",
  learning.rate = 0.001,
  compile = TRUE
)
}
\arguments{
\item{filters}{Integer, the dimensionality of the output space (i.e. the
number of output filters in the convolution).}

\item{kernel_size}{An integer or list of a single integer, specifying the
length of the 1D convolution window.}

\item{residual_blocks}{if a single integer: the number of residual blocks in the
model - the dilation rate of each block `i` is calculated by `2^i`. If a vector,
then it's used as the dilation rate of each residual block.}

\item{input_tensor}{Tensor to use as input for the model. Usually a 3d tensor.}

\item{initial_kernel_size}{kernel size of the first causal convolution.}

\item{initial_filters}{number of filters in the initial causal convolution.}

\item{output_channels}{number of channels in the output.}

\item{output_activation}{activation function for the last layer. Default: 'softmax'}
}
\description{
create wavenet model
}
