% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crispr_full}
\alias{crispr_full}
\title{Example training dataset consiting of a sequence of nucleotides of CRISPR loci
Filtered for unambigous characters and contains only characters in the vocabulary {A,G,G,T
}
contain all CRISPR loci found in NCBI representative genomes using CRT
Can be loaded to workspace via data(crispr_full)}
\format{
Large character of 4.9 Mb
}
\usage{
crispr_full
}
\description{
Example training dataset consiting of a sequence of nucleotides of CRISPR loci
Filtered for unambigous characters and contains only characters in the vocabulary {A,G,G,T
}
contain all CRISPR loci found in NCBI representative genomes using CRT
Can be loaded to workspace via data(crispr_full)
}
\references{
\url{http://github.com/philippmuench}
}
\keyword{datasets}
