% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{deepG-package}
\alias{deepG}
\alias{deepG-package}
\title{deepG for GenomeNet}
\description{
deepG is an open source software library for building deep neural
networks for genomic modeling.
}
\details{
This package generates \href{https://deepg.de}{deepG}

For additional documentation on the deepG package see
\href{https://genomenet.de}{https://genomenet.de}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GenomeNet/deepG}
  \item \url{https://deepg.de/}
  \item Report bugs at \url{https://github.com/GenomeNet/deepG/issues}
}

}
\author{
\strong{Maintainer}: René Mreches \email{rene.mreches@helmholtz-hzi.de}

Authors:
\itemize{
  \item Philipp Münch \email{philipp.muench@helmholtz-hzi.de}
  \item Martin Binder \email{developer.mb706@mb706.com}
  \item Hüseyin Anil Gündüz \email{anil.guenduez@stat.uni-muenchen.de}
  \item Xiao-Yin To \email{x.to@stat.uni-muenchen.de}
  \item Alice McHardy \email{alice.mchardy@helmholtz-hzi.de}
}

}
