% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{deepG}
\alias{deepG}
\title{deepG for GenomeNet}
\description{
deepG is a is an open source software library for building deep neuronal 
networks for genomic modeling
}
\details{
This package generates \href{http://www.genomenet.de}{GenomeNet}

For additional documentation on the deepG package see
\href{https://genomenet.de}{https://genomenet.de}
}
