% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{download_from_ncbi}
\alias{download_from_ncbi}
\title{Download genome from NCBI and returns preprocessed object}
\usage{
download_from_ncbi(accession = "GCF_000007825.1", ...)
}
\arguments{
\item{accession}{refseq specific accession identifier}

\item{verbose}{TRUE/FALSE}
}
\description{
Download genome from NCBI and returns preprocessed object
}
