% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{early_stopping_cb}
\alias{early_stopping_cb}
\title{early stopping callback}
\usage{
early_stopping_cb(early_stopping_patience, early_stopping_time, by_time = TRUE)
}
\arguments{
\item{early_stopping_patience}{Stop training if val_loss does not improve for \code{early_stopping_patience}.}

\item{early_stopping_time}{Time in minutes after which to stop training.}

\item{by_time}{Whether to use time or patience as metric.}
}
\description{
early stopping callback
}
