% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{early_stopping_time_cb}
\alias{early_stopping_time_cb}
\title{Stop training callback}
\usage{
early_stopping_time_cb(stop_time = NULL)
}
\arguments{
\item{stop_time}{Time in seconds after which to stop training.}
}
\value{
A Keras callback that stops training after specified time.
}
\description{
Stop training after specified time.
}
\examples{
\dontshow{if (reticulate::py_module_available("tensorflow")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
est <- early_stopping_time_cb(stop_time = 60)
\dontshow{\}) # examplesIf}
}
