% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli}
\alias{ecoli}
\title{Ecoli genome}
\format{
Large character of 4.64 MB
}
\usage{
data(ecoli)
}
\description{
E. coli genome for evaluation.
Can be loaded to workspace via \code{data(ecoli)}.
}
\references{
\url{https://www.science.org/doi/10.1126/science.277.5331.1453}
}
\keyword{datasets}
