% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_small}
\alias{ecoli_small}
\title{Ecoli subset}
\format{
character 326.73 kB
}
\usage{
data(ecoli_small)
}
\description{
Subset of the E. coli genome for evaluation.
Can be loaded to workspace via \code{data(ecoli_small)}.
}
\references{
\url{https://www.science.org/doi/10.1126/science.277.5331.1453}
}
\keyword{datasets}
