% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorboard.R
\docType{data}
\name{ecoliCustomScalar}
\alias{ecoliCustomScalar}
\title{Custom lambda callback that shows accuacy scores for the E. coli genome in Tensorboard}
\format{
An object of class \code{keras.callbacks.LambdaCallback} (inherits from \code{keras.callbacks.Callback}, \code{python.builtin.object}) of length 2.
}
\usage{
ecoliCustomScalar
}
\description{
Custom lambda callback that shows accuacy scores for the E. coli genome in Tensorboard
}
\keyword{datasets}
