% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate_linear}
\alias{evaluate_linear}
\title{Evaluate matrices of true targets and predictions from layer with linear activation.}
\usage{
evaluate_linear(y_true, y_conf, label_names = NULL)
}
\arguments{
\item{y_true}{Matrix of true labels}

\item{y_conf}{Matrix of predictions.}

\item{label_names}{Names of corresponding labels. Length must be equal to number of columns of \code{y}.}
}
\description{
Compute MAE and MSE, given predictions and
true targets. Outputs columnwise average.
}
\examples{
y_true <- matrix(rnorm(n = 12), ncol = 3)
y_conf <- matrix(rnorm(n = 12), ncol = 3)
evaluate_linear(y_true, y_conf)
}
