% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{evaluateFasta}
\alias{evaluateFasta}
\title{Evaluates a trained model on .fasta/fastq files}
\usage{
evaluateFasta(
  fasta.path,
  model = NULL,
  batch.size = 100,
  step = 1,
  vocabulary = c("a", "c", "g", "t"),
  label_vocabulary = c("a", "c", "g", "t"),
  numberOfBatches = 10,
  filePath = NULL,
  format = "fasta",
  filename = "",
  target_middle = FALSE,
  plot = FALSE,
  mode = "lm",
  acc_per_batch = FALSE,
  output_format = "target_right",
  ambiguous_nuc = "zero",
  evaluate_all_files = FALSE,
  verbose = TRUE,
  max_iter = 20000,
  target_from_csv = NULL
)
}
\arguments{
\item{fasta.path}{Input directory where fasta/fastq files are located.}

\item{model}{A keras model.}

\item{batch.size}{Number of samples per batch.}

\item{step}{How often to take a sample.}

\item{vocabulary}{Vector of allowed characters, character outside vocabulary get encoded as 0-vector.}

\item{label_vocabulary}{Labels for targets. Equal to vocabulary if not given.}

\item{numberOfBatches}{How many batches to evaluate.}

\item{filePath}{Where to store output, if missing output won't be written.}

\item{format}{File format, "fasta" or "fastq".}

\item{filename}{Name of output file.}

\item{plot}{Returns density plot of accuracies if TRUE.}

\item{mode}{Either "lm" for language model and "label_header", "label_csv" or "label_folder" for label classification.}

\item{acc_per_batch}{Whether to return vector with accuracies for every batch.}

\item{output_format}{Determines shape of output tensor for language model.
Either "target_right", "target_middle_lstm", "target_middle_cnn" or "wavenet".
Assume a sequence "AACCGTA". Output correspond as follows
"target_right": X = "AACCGT", Y = "A"
"target_middle_lstm": X = (X_1 = "AAC", X_2 = "ATG"), Y = "C" (note reversed order of X_2)
"target_middle_cnn": X = "AACGTA", Y = "C" 
"wavenet": X = "AACCGT", Y = "ACCGTA"}

\item{ambiguous_nuc}{How to handle nucleotides outside vocabulary, either "zero", "discard" or "equal". If "zero", input gets encoded as zero vector; 
if "equal" input is 1/length(vocabulary) x length(vocabulary). If "discard" samples containing nucleotides outside vocabulary get discarded.}

\item{evaluate_all_files}{Boolean, if TRUE will iterate over all files in \code{fasta.path} once. \code{numberOfBatches} will be overwritten.}

\item{verbose}{Whether to show message.}

\item{max_iter}{Stop after max_iter number of iterations failed to produce a new batch.}

\item{target_from_csv}{Path to csv file with target mapping. One column should be called "file" and other entries in row are the targets.}

\item{model.path}{Path to pretrained model.}
}
\description{
Returns accuracies per batch and overall confusion matrix. Evaluates \code{batch.size} * \code{numberOfBatches} samples.
}
