% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_model}
\alias{example_model}
\title{Trained LSTM model (only for testing)}
\format{keras model}
\usage{
example_model(object)
}
\description{
Trained LSTM model (only for testing)
}
\examples{
head(example_model)
}
\references{
\url{http://github.com/philippmuench}
}
