% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{exp_decay}
\alias{exp_decay}
\title{Exponential Decay}
\usage{
exp_decay(lrmax = 0.005, mult = 0.1, epoch = NULL)
}
\arguments{
\item{lrmax}{Upper limit of the range for the learning rate.}

\item{mult}{Factor, by which the number of epochs until a restart is decreased after a new step.}

\item{epoch}{Epoch, for which the learning rate shall be calculated.}
}
\description{
Compute the learning Rate for a given epoch using Exponential Decay.
}
\examples{
exp_decay(lrmax = 0.005, mult = 0.1, epoch = 8) 

}
