% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{extractCellFromStates}
\alias{extractCellFromStates}
\title{Preparing states for tSNE}
\usage{
extractCellFromStates(
  states.list = NULL,
  states.folder = "example_files/states",
  cell.number = 2,
  flatten = FALSE
)
}
\arguments{
\item{states.list}{list of states generated by generateStatesromFolder}

\item{states.folder}{generat states list from folder containing tate files}

\item{cell.number}{number of cell/neuron that should go into tSNE}

\item{flatten}{if true, all cells will be used}
}
\description{
Preparing states for tSNE
}
