% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{f1_from_conf_matrix}
\alias{f1_from_conf_matrix}
\title{Compute F1-score from confusion matrix.}
\usage{
f1_from_conf_matrix(conf_mat)
}
\arguments{
\item{conf_mat}{A confusion matrix. Columns should 
correspond to true label and rows to predictions.}
}
\description{
Compute F1-score from confusion matrix.
}
