% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{focal_loss_multiclass}
\alias{focal_loss_multiclass}
\title{Focal loss for two or more labels}
\usage{
focal_loss_multiclass(y_true, y_pred, gamma = 2, alpha = 0.25)
}
\description{
Focal loss for two or more labels
}
