% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{function_args_cb}
\alias{function_args_cb}
\title{function arguments callback}
\usage{
function_args_cb(argumentList, tensorboard.log, run.name)
}
\arguments{
\item{tensorboard.log}{Path to tensorboard log directory.}

\item{run.name}{Name of the run (without file ending). Name will be used to identify output from callbacks.}
}
\description{
function arguments callback
}
