% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{gen_rds}
\alias{gen_rds}
\title{Read batches from rds arrays.}
\usage{
gen_rds(
  rds_folder,
  batch_size,
  fileLog = NULL,
  max_samples = NULL,
  proportion_per_file = NULL,
  target_len = NULL,
  n_gram = NULL,
  n_gram_stride = 1
)
}
\arguments{
\item{rds_folder}{Path to input files.}
}
\description{
Read batches from rds arrays.
}
