% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{generateStatesromFolder}
\alias{generateStatesromFolder}
\title{generate states list for tSNE from a fasta folder
it is expected that the length of each fasta file is equal}
\usage{
generateStatesromFolder(fasta.folder = "example_files/fasta",
  model.path = "example_files/dummy_model_cpu.hdf5")
}
\arguments{
\item{fasta.folder}{path to fasta folder}

\item{model.path}{path to model file}

\item{cell.numer}{number of cell to be extracted}

\item{flatten}{if ture, all cell number will be used}
}
\description{
generate states list for tSNE from a fasta folder
it is expected that the length of each fasta file is equal
}
