% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generator_dummy}
\alias{generator_dummy}
\title{Random data generator}
\usage{
generator_dummy(model, batch_size)
}
\arguments{
\item{model}{A keras model.}
}
\description{
Creates a random input/target list once and repeatedly returns list.
}
