% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generator_fasta_label_folder_wrapper}
\alias{generator_fasta_label_folder_wrapper}
\title{Generator wrapper}
\usage{
generator_fasta_label_folder_wrapper(
  val,
  new_batch_size = NULL,
  samples_per_target = NULL,
  batch_size = NULL,
  path = NULL,
  voc_len = NULL,
  maxlen = NULL,
  reshape_mode = NULL,
  buffer_len = NULL,
  concat_maxlen = NULL
)
}
\arguments{
\item{val}{Train or validation generator.}

\item{path}{Path}
}
\description{

}
