% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_class_weight}
\alias{get_class_weight}
\title{Estimate frequency of different classes}
\usage{
get_class_weight(
  path,
  vocabulary_label = NULL,
  format = "fasta",
  file_proportion = 1,
  train_type = "label_folder",
  csv_path = NULL
)
}
\arguments{
\item{vocabulary_label}{Character vector of possible targets. Targets outside \code{vocabulary_label} will get discarded.}

\item{format}{File format, either "fasta" or "fastq".}

\item{file_proportion}{Proportion of files to randomly sample for estimating class distributions.}

\item{train_type}{Either "label_folder", "label_header" or "label_csv".}
}
\description{
Count number of nucleotides for each class and use as estimation for relation of class distribution.
Outputs list of class relations. Can be used as input for \code{class_weigth} in \code{\link{train_model}} function.
}
