% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{get_hyper_param}
\alias{get_hyper_param}
\title{Extract hyperparameters from model}
\usage{
get_hyper_param(model)
}
\arguments{
\item{model}{A keras model.}
}
\value{
List of hyperparameters.
}
\description{
Extract hyperparameters from model
}
\keyword{internal}
