% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{get_optimizer}
\alias{get_optimizer}
\title{Get solver and learning_rate from model.}
\usage{
get_optimizer(model)
}
\value{
Keras optimizer.
}
\description{
Get solver and learning_rate from model.
}
\keyword{internal}
