% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_utils.R
\name{get_output_activations}
\alias{get_output_activations}
\title{Get activation functions of output layers}
\usage{
get_output_activations(model)
}
\arguments{
\item{model}{A keras model.}
}
\value{
Character vector with names of activation functions of model outputs.
}
\description{
Get activation functions of output layers.
}
\examples{
model <-  create_model_lstm_cnn(
  maxlen = 50,
  layer_lstm = 8,
  layer_dense = c(64, 2),
  verbose = FALSE)
get_output_activations(model)

}
