% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_run_name}
\alias{get_run_name}
\title{Generate run_name if none is given or is already present.}
\usage{
get_run_name(
  run_name = NULL,
  path_tensorboard,
  path_checkpoint,
  path_log,
  auto_extend = FALSE
)
}
\arguments{
\item{auto_extend}{If run_name is already present, add "_2" to name. If name already ends with "_x" replace x with x+1.}
}
\description{
Generate run_name if none is given or is already present.
}
\keyword{internal}
