% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_vocabulary}
\alias{get_vocabulary}
\title{Returns the vocabulary from character string}
\usage{
get_vocabulary(char, verbose = F)
}
\arguments{
\item{char}{a character string of text, length of one}

\item{verbose}{TRUE/FALSE}
}
\description{
Use this function with a character string as input such as data(train)
}
