% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{getEpochLR}
\alias{getEpochLR}
\title{Learning Rate Calculator}
\usage{
getEpochLR(lr_schedule, epoch)
}
\arguments{
\item{lr_schedule}{The name of a learning rate schedule.}

\item{epoch}{Epoch, for which the learning rate shall be calculated.}
}
\description{
Computes the learning rate for a given epoch.
}
\keyword{internal}
