% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{getStates}
\alias{getStates}
\title{Get cell states of semi-redundant chunks}
\usage{
getStates(
  model.path,
  x,
  maxlen = 30,
  batch.size = 100,
  run.name = "output",
  type = "csv",
  verbose = F
)
}
\arguments{
\item{model.path}{path to keras model in hdf5 format}

\item{x}{semi-redundant chunks (one-hot)}

\item{maxlen}{time steps to unroll for}

\item{batch.size}{how many samples are trained in parallel}

\item{run.name}{name of output files without ending}

\item{type}{save-type, will save as hdf5 if type is set to 'hdf5' (default .csv)}

\item{verbose}{TRUE/FALSE}
}
\description{
Get cell states of semi-redundant chunks
}
