% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-states.R
\name{getstates_ncbi}
\alias{getstates_ncbi}
\title{Iterates over entries in NCBI list and export cell state responses from a model}
\usage{
getstates_ncbi(list_path, model_path, ...)
}
\arguments{
\item{list_path}{path to file downloaded from NCBI}
}
\description{
Iterates over entries in NCBI list and export cell state responses from a model
}
