% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gradients_stepwise}
\alias{gradients_stepwise}
\title{Compute gradients stepwise (one batch at a time)}
\usage{
gradients_stepwise(
  model = model,
  baseline_seq,
  target_class_idx,
  input_idx = NULL
)
}
\description{
Compute gradients stepwise (one batch at a time)
}
