% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmaps_integrated_grad}
\alias{heatmaps_integrated_grad}
\title{Creates a heatmap from output of \code{integrated_gradients} function.}
\usage{
heatmaps_integrated_grad(integrated_grads, input_seq)
}
\arguments{
\item{integrated_grads}{Output of \code{integrated_gradients} function}

\item{input_seq}{Input sequence for model. Should be the same as \code{input_seq} input for corresponding
\code{integrated_gradients} call that computed \code{integrated_grads} argument.}
}
\description{
Creates a heatmap from output of \code{integrated_gradients} function.
}
