% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{hyper_param_model_outside_cb}
\alias{hyper_param_model_outside_cb}
\title{Hyperparameter callback}
\usage{
hyper_param_model_outside_cb(
  path_tensorboard,
  run_name,
  wavenet_format,
  cnn_format,
  model,
  vocabulary,
  path,
  reverse_complement,
  vocabulary_label,
  maxlen,
  epochs,
  max_queue_size,
  lr_plateau_factor,
  batch_size,
  patience,
  cooldown,
  steps_per_epoch,
  step,
  shuffle_file_order
)
}
\arguments{
\item{path_tensorboard}{Path to path_tensorboard directory or NULL. If NULL, training not tracked on tensorboard.}

\item{run_name}{Name of the run. Name will be used to identify output from callbacks.}

\item{model}{A keras model.}

\item{vocabulary}{Vector of allowed characters. Character outside vocabulary get encoded as specified in \code{ambiguous_nuc}.}

\item{path}{Path to folder where individual or multiple FASTA or FASTQ files are located for training. If \code{train_type} is \code{label_folder}, should be a vector or list
where each entry corresponds to a class. If \code{train_type} is not \code{label_folder}, can be a list of directories and/or single files.}

\item{reverse_complement}{Logical, for every new file decide randomly to use original data or its reverse complement.}

\item{vocabulary_label}{Character vector of possible targets. Targets outside \code{vocabulary_label} will get discarded if
\code{train_type = "label_header"}.}

\item{epochs}{Number of iterations.}

\item{max_queue_size}{Queue on fit_generator().}

\item{lr_plateau_factor}{Factor of decreasing learning_rate when plateau is reached.}

\item{batch_size}{Number of samples used for one network update.}

\item{patience}{Number of epochs waiting for decrease in val_loss before reducing learning_rate.}

\item{cooldown}{Number of epochs without changing learning_rate.}

\item{steps_per_epoch}{Number of training batches per epoch.}

\item{step}{Frequency of sampling steps.}

\item{shuffle_file_order}{Boolean, whether to go through files sequentially or shuffle beforehand.}
}
\description{
Hyperparameter callback
}
