% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{integrated_gradients}
\alias{integrated_gradients}
\title{Computes integrated gradients (https://www.tensorflow.org/tutorials/distribute/keras) for model and an input sequence. 
This can be used to visiulize what part of the input is import for the models decision.}
\usage{
integrated_gradients(
  m_steps = 50,
  baseline_type = "zero",
  input_seq,
  target_class_idx,
  model,
  pred_stepwise = FALSE,
  num_baseline_repeats = 1
)
}
\arguments{
\item{m_steps}{Number of steps between baseline and original input.}

\item{baseline_type}{Baseline sequence, either "zero" for all zeros or "shuffle" for random permutation of input_seq.}

\item{input_seq}{Input tensor.}

\item{target_class_idx}{Index of class to compute gradient for}

\item{model}{Model to compute gradient for.}

\item{pred_stepwise}{Whether to do predictions with batch size 1 rather than all at once. Can be used if
input is too big to handle at once. Only supported for single input layer.}

\item{num_baseline_repeats}{Number of different baseline estimations if baseline_type is "shuffle" (estimate integrated 
gradient repeatedly for different shufflings). Final result is average of \code{num_baseline} single calculations.}
}
\description{
Computes integrated gradients (https://www.tensorflow.org/tutorials/distribute/keras) for model and an input sequence. 
This can be used to visiulize what part of the input is import for the models decision.
}
