% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{interpolate_seq}
\alias{interpolate_seq}
\title{Interpolation between baseline and prediction}
\usage{
interpolate_seq(m_steps = 50, baseline_type = "shuffle", input_seq)
}
\arguments{
\item{m_steps}{Number of steps between baseline and original input.}

\item{baseline_type}{Baseline sequence, either "zero" for all zeros or "shuffle" for random permutation of input_seq.}

\item{input_seq}{Input tensor.}
}
\description{
Interpolation between baseline and prediction
}
\keyword{internal}
