% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{labelByFolderGeneratorWrapper}
\alias{labelByFolderGeneratorWrapper}
\title{Generator wrapper}
\usage{
labelByFolderGeneratorWrapper(val, path)
}
\arguments{
\item{val}{Train or validation generator.}

\item{path}{Path}
}
\description{

}
