% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{layer_add_time_dist_wrapper}
\alias{layer_add_time_dist_wrapper}
\title{Add layer}
\usage{
layer_add_time_dist_wrapper(load_r6 = FALSE)
}
\arguments{
\item{load_r6}{Whether to load the R6 layer class.}
}
\description{
Add output of time distribution representations.
}
