% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{layer_aggregate_time_dist_wrapper}
\alias{layer_aggregate_time_dist_wrapper}
\title{Add layer}
\usage{
layer_aggregate_time_dist_wrapper(load_r6 = FALSE, method = "sum")
}
\arguments{
\item{load_r6}{Whether to load the R6 layer class.}

\item{method}{At least one of the options, \verb{"sum", "max"} or \code{"mean"}.
l <- layer_aggregate_time_dist_wrapper()}
}
\description{
Add output of time distribution representations.
}
