% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{load_cp}
\alias{load_cp}
\title{Load checkpoint}
\usage{
load_cp(
  cp_path,
  cp_filter = NULL,
  ep_index = NULL,
  compile = FALSE,
  learning_rate = 0.01,
  solver = "adam",
  loss = "categorical_crossentropy",
  verbose = TRUE
)
}
\arguments{
\item{cp_path}{A directory containing checkpoints or a single checkpoint file.
If a directory, choose checkpoint based on \code{cp_filter} or \code{ep_index}.}

\item{cp_filter}{Condition to choose checkpoint if \code{cp_path} is a directory.
Either "acc" for bast validation accuracy, "loss" for best validation loss or "last_ep" for last epoch.}

\item{ep_index}{Load checkpoint from specific epoch number.}

\item{compile}{Whether to compile the model.}

\item{learning_rate}{Learning rate for optimizer.}

\item{solver}{Optimization method, options are \verb{"adam", "adagrad", "rmsprop"} or \code{"sgd"}.}

\item{loss}{Loss function. Only used if model gets compiled.}

\item{verbose}{Whether to print choosen checkpoint path.}
}
\description{
Load checkpoint from directory. Chooses best checkpoint based on some condition. Condition
can be best accuracy, best loss, last epoch number or a specified epoch number.
}
