% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{load_model_self_genomenet}
\alias{load_model_self_genomenet}
\title{Load pretrained self-genomenet model}
\usage{
load_model_self_genomenet(
  maxlen,
  compile = FALSE,
  optimizer = "adam",
  learning_rate = NULL
)
}
\arguments{
\item{maxlen}{Model input size. Either 150 or 10000.}

\item{compile}{Whether to compile the model.}

\item{learning_rate}{Learning rate for optimizer. If compile is TRUE and learning_rate is NULL,
will use learning rate from previous training.}
}
\description{
Classification model with labels "bacteria", "virus-no-phage","virus-phage".
TODO: add link to paper
}
