% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{log_cb}
\alias{log_cb}
\title{Log callback}
\usage{
log_cb(path_log, run_name)
}
\arguments{
\item{path_log}{Path to output directory.}

\item{run_name}{Name of output file is run_name + ".csv".}
}
\description{
Log callback
}
\keyword{internal}
