% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{loss_cl}
\alias{loss_cl}
\title{Contrastive loss}
\usage{
loss_cl(margin = 1)
}
\arguments{
\item{margin}{Integer, baseline for distance for which pairs should be classified as dissimilar.}
}
\description{
Contrastive loss as used here: https://keras.io/examples/vision/siamese_contrastive/.
}
