% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{LRstop}
\alias{LRstop}
\title{Learning Rate Schedule - Parameter Check}
\usage{
LRstop(lr_schedule)
}
\arguments{
\item{lr_schedule}{The name of a learning rate schedule.}
}
\description{
Checks, whether all necessary parameters for a defined learning rate schedule are given.
}
\keyword{internal}
