% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{merge_models}
\alias{merge_models}
\title{Merge two models}
\usage{
merge_models(
  models,
  layer_names,
  layer_dense,
  solver = "adam",
  learning_rate = 1e-04,
  freeze_base_model = c(FALSE, FALSE)
)
}
\arguments{
\item{models}{List of two models}

\item{layer_names}{Vector of length 2 with names of layers to merge.}

\item{layer_dense}{Dense layers of size layer_dense after last LSTM (or last CNN is \code{layers.lstm = 0}) layer.}

\item{solver}{Optimization method, options are "adam", "adagrad", "rmsprop" or "sgd".}

\item{learning_rate}{Learning rate for optimizer.}
}
\description{
Merge two models
}
