% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{messagef}
\alias{messagef}
\title{Wrapper for message(sprintf)}
\usage{
messagef(..., newline = TRUE)
}
\arguments{
\item{...}{text input}

\item{newline}{print in new line}
}
\description{
Wrapper for message(sprintf)
}
