% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{model_card_cb}
\alias{model_card_cb}
\title{Create model card}
\usage{
model_card_cb(model_card_path = NULL, run_name, argumentList)
}
\arguments{
\item{model_card_path}{Directory for model card logs.}
}
\description{
Log information about model, hyperparameters, generator options, training data, scores etc
}
