% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_self_genomenet_maxlen_10k}
\alias{model_self_genomenet_maxlen_10k}
\title{Genomenet model maxlen 10000}
\format{
raw
}
\usage{
model_self_genomenet_maxlen_10k
}
\description{
Serialized keras model for "bacteria", "virus-no-phage","virus-phage" classification.
Can be loaded to workspace via data(self_genomenet_model_maxlen_10k)
Use keras::unserialize_model to transform to keras model or load with
load_model_self_genomenet(maxlen = 10000) function call.
}
\keyword{datasets}
