% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_self_genomenet_maxlen_150}
\alias{model_self_genomenet_maxlen_150}
\title{Serialized keras model for "bacteria", "virus-no-phage","virus-phage" classification.
Can be loaded to workspace via data(self_genomenet_model_maxlen_150).
Use keras::unserialize_model to transform to keras model or load with
load_model_self_genomenet(maxlen = 150) function call.}
\format{
raw
}
\usage{
model_self_genomenet_maxlen_150
}
\description{
Serialized keras model for "bacteria", "virus-no-phage","virus-phage" classification.
Can be loaded to workspace via data(self_genomenet_model_maxlen_150).
Use keras::unserialize_model to transform to keras model or load with
load_model_self_genomenet(maxlen = 150) function call.
}
\keyword{datasets}
