% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{modelstep}
\alias{modelstep}
\title{Step function}
\usage{
modelstep(trainvaldat, model, train_type = "cpc", training = F)
}
\arguments{
\item{trainvaldat}{A data generator.}

\item{model}{A keras model.}

\item{train_type}{Either \code{"cpc"}, \code{"Self-GenomeNet"}.}

\item{training}{Boolean. Whether this step is a training step.}
}
\description{
Step function
}
\keyword{internal}
