% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_gram.R
\name{n_gram}
\alias{n_gram}
\title{Get distribution of n-grams}
\usage{
n_gram(
  path_input,
  n = 2,
  vocabulary = c("A", "C", "G", "T"),
  format = "fasta",
  file_sample = NULL,
  step = 1,
  nuc_dist = FALSE
)
}
\arguments{
\item{path_input}{Path to folder containing fasta files or single fasta file.}

\item{n}{Size of n gram.}

\item{vocabulary}{Vector of allowed characters, samples outside vocabulary get discarded.}

\item{format}{"fasta" or "fastq".}

\item{file_sample}{If integer, size of random sample of files in \code{path_input}.}

\item{step}{Frequency of samples.}
}
\value{
Returns a matrix with distributions of nucleotides given the previous n nucleotides.
}
\description{
Get distribution of n-grams
}
