% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{noisy_loss_wrapper}
\alias{noisy_loss_wrapper}
\title{Loss function for label noise}
\usage{
noisy_loss_wrapper(noise_matrix)
}
\arguments{
\item{noise_matrix}{Matrix of noise distribution.
If first label contains 5% wrong labels and second label no noise, then
\code{noise_matrix <- matrix(c(0.95, 0.05, 0, 1), nrow = 2, byrow = TRUE)}.}
}
\description{
Implements approach from https://arxiv.org/abs/1609.03683 and code from 
https://github.com/giorgiop/loss-correction/blob/15a79de3c67c31907733392085c333547c2f2b16/loss.py#L16-L21
}
