% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{nopatchescalc}
\alias{nopatchescalc}
\title{Number of Patches calculation}
\usage{
nopatchescalc(plen, maxlen, stride)
}
\arguments{
\item{plen}{Length of a patch.}

\item{maxlen}{Length of the input sequence.}

\item{stride}{Stride.}
}
\description{
Compute the Number of Patches.
}
\keyword{internal}
