% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_viewer.R
\name{nucleotide2value}
\alias{nucleotide2value}
\title{Change nucleotides to values}
\usage{
nucleotide2value(nucleotides)
}
\arguments{
\item{nucleotides}{nucleotides from the genomic sequence}
}
\description{
Change nucleotides to values
}
