% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{one_beam}
\alias{one_beam}
\title{One beam}
\usage{
one_beam(sequence, model, vocabulary, num_alternatives = 2)
}
\arguments{
\item{sequence}{input sequence}

\item{model}{trained lstm}

\item{vocabulary}{ordered vocabulary of input sequence}

\item{num_alternatives}{number of best solutions}
}
\description{
One beam
}
