% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_model}
\alias{predict_model}
\title{Make prediction for nucleotide sequence or entries in fasta/fastq file}
\usage{
predict_model(
  output_format = "one_seq",
  model = NULL,
  layer_name = NULL,
  sequence = NULL,
  path_input = NULL,
  round_digits = 2,
  filename = "states.h5",
  step = 1,
  vocabulary = c("a", "c", "g", "t"),
  batch_size = 256,
  verbose = TRUE,
  return_states = FALSE,
  padding = FALSE,
  output_type = "h5",
  path_model = NULL,
  mode = "lm",
  lm_format = "target_right",
  output_dir = NULL,
  format = "fasta",
  include_seq = TRUE,
  reverse_complement_encoding = FALSE
)
}
\arguments{
\item{output_format}{Either "one_seq", "by_entry", "by_entry_one_file", "one_pred_per_entry".
If "one_seq" computes prediction for sequence argument or fasta/fastq file.
Combines fasta entries in file to one sequence. This means predictor sequences can contain elements from more than one fasta entry.
If "by_entry" will output a separate file for each fasta/fastq entry.
Names of output files are: output_dir + "Nr" + i + filename + output_type, where i is the number of the fasta entry.
If "by_entry_one_file", will store prediction for all fasta entry in one h5 file.
If "one_pred_per_entry" will make one prediction for each entry by either picking random sample for long sequences
or pad sequence for short sequences.}

\item{model}{A keras model. If model and path_model are not NULL, model will be used for inference.}

\item{layer_name}{Name of layer to get output from. If NULL, will use the last layer.}

\item{sequence}{Character string, ignores path_input if argument given.}

\item{path_input}{Path to fasta file.}

\item{round_digits}{Number of decimal places.}

\item{filename}{Filename to store states in. No file output if argument is NULL.}

\item{step}{Frequency of sampling steps.}

\item{vocabulary}{Vector of allowed characters, character outside vocabulary get encoded as 0-vector.}

\item{batch_size}{Number of samples to evaluate at once. Does not change output, only relevant for speed and memory.}

\item{verbose}{Whether to print model before and after removing layers.}

\item{return_states}{Return predictions as data frame. Only supported for output_format "one_seq".}

\item{padding}{Logical scalar, generate states for first maxlen nucleotides by
padding beginning of sequence with 0-vectors.}

\item{output_type}{"h5" or "csv". If output_format is "by_entries_one_file", "one_pred_per_entry" can only be "h5".}

\item{path_model}{Path to a pretrained model.}

\item{mode}{Either "lm" for language model or "label" for label classification.}

\item{output_dir}{Path to folder, where to write output.}

\item{format}{Either "fasta" or "fastq".}

\item{include_seq}{Whether to include input sequence in h5 file.}
}
\description{
Removes layers (optional) from pretrained model and calculates states of fasta/fastq file or nucleotide sequence.
Writes states to h5/csv file. To acces the content of h5 output use \code{load_predictions} function.
}
