% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R, R/predict_2.R
\name{predict_model_by_entry}
\alias{predict_model_by_entry}
\title{Write states to h5 file}
\usage{
predict_model_by_entry(
  path_model = NULL,
  layer_name = NULL,
  path_input,
  round_digits = 2,
  filename = "states.h5",
  output_dir,
  step = 1,
  vocabulary = c("a", "c", "g", "t"),
  batch_size = 256,
  padding_maxlen = FALSE,
  output_type = "h5",
  model = NULL,
  mode = "lm",
  lm_format = "target_right",
  format = "fasta",
  reverse_complement_encoding = FALSE,
  verbose = FALSE,
  include_seq = FALSE
)

predict_model_by_entry(
  path_model = NULL,
  layer_name = NULL,
  path_input,
  round_digits = 2,
  filename = "states.h5",
  output_dir,
  step = 1,
  vocabulary = c("a", "c", "g", "t"),
  batch_size = 256,
  padding_maxlen = FALSE,
  output_type = "h5",
  model = NULL,
  mode = "lm",
  lm_format = "target_right",
  format = "fasta",
  reverse_complement_encoding = FALSE,
  verbose = FALSE,
  include_seq = FALSE
)
}
\arguments{
\item{filename}{Filename to store states, function adds "\emph{nr}" + "i" after name, where i is entry number.}

\item{output_dir}{Path to folder, where to write output.}
}
\description{
Removes layers (optional) from pretrained model and calculates states of fasta file, writes a separate
h5 file for every fasta entry in fasta file. h5 files also contain the nucleotide sequence and positions of targets corresponding to states.
Names of output files are: file_path + "Nr" + i + filename + output_type, where i is the number of the fasta entry.

Removes layers (optional) from pretrained model and calculates states of fasta file, writes a separate
h5 file for every fasta entry in fasta file. h5 files also contain the nucleotide sequence and positions of targets corresponding to states.
Names of output files are: file_path + "Nr" + i + filename + output_type, where i is the number of the fasta entry.
}
\keyword{internal}
