% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_model_by_entry_one_file}
\alias{predict_model_by_entry_one_file}
\title{Write states to h5 file}
\usage{
predict_model_by_entry_one_file(
  path_model,
  path_input,
  round_digits = 2,
  filename = "states.h5",
  step = 1,
  vocabulary = c("a", "c", "g", "t"),
  batch_size = 256,
  layer_name = NULL,
  padding = FALSE,
  verbose = TRUE,
  model = NULL,
  mode = "lm",
  target_middle = FALSE,
  format = "fasta",
  include_seq = TRUE,
  reverse_complement_encoding = FALSE
)
}
\description{
Removes layers (optional) from pretrained model and calculates states of fasta file,
writes separate states matrix in one .h5 file for every fasta entry.
h5 file also contains the nucleotide sequences and positions of targets corresponding to states.
}
