% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_model_one_seq}
\alias{predict_model_one_seq}
\title{Write output of specific model layer to h5 or csv file.}
\usage{
predict_model_one_seq(
  path_model = NULL,
  layer_name = NULL,
  sequence = NULL,
  path_input = NULL,
  round_digits = 2,
  filename = "states.h5",
  step = 1,
  vocabulary = c("a", "c", "g", "t"),
  batch_size = 256,
  verbose = TRUE,
  return_states = FALSE,
  target_len = 1,
  output_type = "h5",
  model = NULL,
  mode = "lm",
  lm_format = "target_right",
  ambiguous_nuc = "zero",
  padding = "none",
  format = "fasta",
  output_dir = NULL,
  include_seq = TRUE,
  reverse_complement_encoding = FALSE,
  ...
)
}
\arguments{
\item{path_model}{Path to a pretrained model.}

\item{layer_name}{Name of layer to get output from. If \code{NULL}, will use the last layer.}

\item{sequence}{Character string, ignores path_input if argument given.}

\item{path_input}{Path to fasta file.}

\item{round_digits}{Number of decimal places.}

\item{filename}{Filename to store states in. No file output if argument is \code{NULL}.}

\item{step}{Frequency of sampling steps.}

\item{vocabulary}{Vector of allowed characters, character outside vocabulary get encoded as 0-vector.}

\item{batch_size}{Number of samples to evaluate at once. Does not change output, only relevant for speed and memory.}

\item{verbose}{Whether to print model before and after removing layers.}

\item{return_states}{Logical scalar, return states matrix.}

\item{target_len}{Number of nucleotides to predict at once for language model.}

\item{output_type}{Either \code{"h5"} or \code{"csv"}.}

\item{model}{A keras model. If model and path_model are not NULL, model will be used for inference.}

\item{mode}{Either \code{"lm"} for language model or \code{"label"} for label classification.}

\item{ambiguous_nuc}{\code{"zero"} or \code{"equal"}.}

\item{padding}{Whether to pad sequences too short for one sample with zeros.}

\item{format}{Either \code{"fasta"} or \code{"fastq"}.}

\item{include_seq}{Whether to include input sequence in h5 file.}

\item{...}{Further arguments for sequence encoding with \code{\link{seq_encoding_label}}.}
}
\description{
Removes layers (optional) from pretrained model and calculates states of fasta file, writes states to h5/csv file.
Function combines fasta entries in file to one sequence. This means predictor sequences can contain elements from more than one fasta entry.
h5 file also contains sequence and positions of targets corresponding to states.
}
\keyword{internal}
