% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_next_nucleotide}
\alias{predict_next_nucleotide}
\title{Predict next nucleotides in sequence output as S4 class}
\usage{
predict_next_nucleotide(sequence, model, vocabulary = c("\\n", "a", "c",
  "g", "t"))
}
\arguments{
\item{sequence}{input sequence}

\item{model}{trained lstm}

\item{vocabulary}{ordered vocabulary of input sequence}
}
\description{
Predict next nucleotides in sequence output as S4 class
}
