% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_gram.R
\name{predict_with_n_gram}
\alias{predict_with_n_gram}
\title{predict the next nucleotide using n-gram}
\usage{
predict_with_n_gram(
  fasta_path,
  distribution_matrix,
  default_pred = "random",
  vocabulary = c("A", "C", "G", "T"),
  file_sample = NULL,
  format = "fasta",
  return_data_frames = FALSE
)
}
\arguments{
\item{fasta_path}{Path to folder containing fasta files or single fasta file.}

\item{distribution_matrix}{Output of \code{n_gram} function.}

\item{default_pred}{Either character from vocabulary or "random". Will be used as prediction if certain n-gram did not appear before.
If "random" assign random prediction.}

\item{vocabulary}{Vector of allowed characters, samples outside vocabulary get discarded.}

\item{file_sample}{If integer, size of random sample of files in \code{fasta_path}.}

\item{format}{"fasta" or "fastq".}

\item{return_data_frames}{Boolean, whether to return data frame with input, predictions, target position and true target.}
}
\description{
predict the next nucleotide using n-gram
}
\examples{
\dontrun{
n_gram_dist <- n_gram(fasta_path = "/net/sgi/genomenet/data/ncbi_genomes/train",
                      n = 2,
                      file_sample = NULL,
                      nuc_dist = TRUE)

# use distribution matrix to make predictions for one file
predictions <- predict_with_n_gram(fasta_path = "/net/sgi/genomenet/data/ncbi_genomes/test/GCF_001941465.1_ASM194146v1_genomic.fasta",
                                   distribution_matrix = n_gram_dist)

# show accuracy
predictions$accuracy
}
}
