% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{prediction-class}
\alias{prediction-class}
\alias{prediction}
\title{An S4 class for the prediction}
\description{
An S4 class for the prediction
}
\section{Slots}{

\describe{
\item{\code{next_char}}{Next predicted char}

\item{\code{probability}}{probability for the predicted char}

\item{\code{index}}{index}

\item{\code{alternative_probability}}{all probabilities for the possible chars}

\item{\code{solution}}{sequence with the predicted char}
}}

