% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictNextNucleotide}
\alias{predictNextNucleotide}
\title{Predict next nucleotides in sequence}
\usage{
predictNextNucleotide(
  sequence,
  model,
  vocabulary = c("l", "a", "c", "g", "t"),
  verbose = F
)
}
\arguments{
\item{sequence}{input sequence, length should be in sync with the model.
If length exceeds input.shape of model then only the right side of the
sequence will be used.}

\item{model}{trained model from the function \code{trainNetwork()}}

\item{vocabulary}{vocabulary of input sequence}

\item{verbose}{TRUE/FALSE}
}
\description{
The output is a S4 class.
}
\examples{
\dontrun{
example.model <- keras::load_model_hdf5("example_model.hdf5")
sequence <- strrep("A", 100)
predictNextNucleotide(sequence, example.model)}
}
