% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{predReads}
\alias{predReads}
\title{Get states for label classification model}
\usage{
predReads(
  model.path = NULL,
  layer.depth = NULL,
  layer_name = NULL,
  fasta.path,
  round_digits = 2,
  format = "fasta",
  filename = "states.h5",
  vocabulary = c("a", "c", "g", "t"),
  batch.size = 256,
  verbose = TRUE,
  returnStates = FALSE,
  model = NULL
)
}
\arguments{
\item{model.path}{Path to a pretrained model.}

\item{layer.depth}{Depth of layer to evaluate. If NULL last layer is used. Can only be used for sequential model. 
Use \code{layer_name} for non-sequential model-}

\item{layer_name}{Name of layer to get output from.}

\item{fasta.path}{Path to fasta file.}

\item{round_digits}{Number of decimal places.}

\item{format}{Either "fasta" or "fastq".}

\item{filename}{Filename to store states in.}

\item{vocabulary}{Vector of allowed characters, character outside vocabulary get encoded as 0-vector.}

\item{batch.size}{Number of samples to evaluate at once. Does not change output, only relevant for speed and memory.}

\item{verbose}{Whether to print model before and after removing layers.}

\item{returnStates}{Logical scalar, return states matrix.}

\item{model}{A keras model. If model and model.path are not NULL, model will be used for inference.}
}
\description{
Computes output at certain model layer. Forces every fasta entry to have length maxlen by either padding sequences shorter than maxlen or taking random subsequence for
longer sequences.
}
