% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_fasta}
\alias{preprocess_fasta}
\title{preprocess of one fasta file}
\usage{
preprocess_fasta(path, maxlen = 30, vocabulary = c("\\n", "a", "c",
  "g", "t"), verbose = F)
}
\description{
preprocess of one fasta file
}
