% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessSemiRedundant}
\alias{preprocessSemiRedundant}
\title{Preprocess string to semi-redundant one-hot vector}
\usage{
preprocessSemiRedundant(
  char,
  maxlen = 250,
  vocabulary = c("l", "p", "a", "c", "g", "t"),
  verbose = F
)
}
\arguments{
\item{char}{character input string of text with the length of one}

\item{maxlen}{length of the semi-redundant sequences}

\item{vocabulary}{char contains the vocabulary from the input char
If no vocabulary exists, it is generated from the input char}

\item{verbose}{TRUE/FALSE}
}
\description{
Outputs semi-redundant set of input character string.
Collapse, tokenize, and vectorize the character.
Use this function with a character string as input. For example, 
if the input text is ABCDEFGHI and the length(maxlen) is 5, the generating chunks would be:
X(1): ABCDE and Y(1): F;
X(2): BCDEF and Y(2): G;
X(3): CDEFG and Y(3): H;
X(4): DEFGH and Y(4): I
}
