% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{quality_to_probability}
\alias{quality_to_probability}
\title{convert fastq quality score to probability}
\usage{
quality_to_probability(quality_vector)
}
\description{
convert fastq quality score to probability
}
