% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{random_subset}
\alias{random_subset}
\title{Take random subset}
\usage{
random_subset(
  proportion_per_file,
  fasta.file,
  use_beta_dist = FALSE,
  use_quality_score
)
}
\description{
Take random subset
}
