% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{ReadOpt}
\alias{ReadOpt}
\title{Reading Pretrained Model function}
\usage{
ReadOpt(pretrained_model)
}
\arguments{
\item{pretrained_model}{The path to a saved keras model.}
}
\description{
Reading Pretrained Model function
}
\keyword{internal}
