% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{readRowsFromH5}
\alias{readRowsFromH5}
\title{Read states from h5}
\usage{
readRowsFromH5(
  h5_path,
  rows,
  verbose = TRUE,
  complete = FALSE,
  getTargetPositions = FALSE
)
}
\arguments{
\item{h5_path}{Path to h5 file.}

\item{rows}{Range of rows to read.}

\item{complete}{Return all entries if TRUE.}

\item{getTargetPositions}{Return position of corresponding targets if TRUE.}
}
\description{

}
