% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{remove_add_layers}
\alias{remove_add_layers}
\title{Remove layers from model and add dense layers}
\usage{
remove_add_layers(
  model = NULL,
  layer_name = NULL,
  layer_depth = NULL,
  dense_layers = NULL,
  last_activation = list("softmax"),
  output_names = NULL,
  losses = NULL,
  verbose = TRUE,
  dropout = NULL,
  freeze_base_model = FALSE,
  compile = FALSE,
  lr = 0.001,
  solver = "adam"
)
}
\arguments{
\item{model}{A keras model. If model and model_path are both not NULL, model_path will be used.}

\item{layer_name}{Name of last layer to use from old model.}

\item{dense_layers}{List of vectors specifiying number of units for each dense layer. If this is a list of length > 1, model 
has multiple output layers.}

\item{last_activation}{List of activations for last entry for each list entry from \code{dense_layers}. Either "softmax" or "sigmoid".}

\item{output_names}{List of names for each output layer.}

\item{losses}{List of loss function for each output.}

\item{verbose}{Boolean.}

\item{dropout}{List of vectors with dropout rates for each new dense layer.}

\item{freeze_base_model}{Whether to freeze all weights before new dense layers.}

\item{compile}{Boolean, whether the new model is compiled or not}

\item{lr}{learning rate if compile == TRUE, default -> learning rate of the old model}

\item{solver}{"adam", "adagrad", "rmsprop" or "sgd" if compile == TRUE, default -> solver of the old model}
}
\description{
Remove layers from model and add dense layers
}
