% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{replace_char_beam}
\alias{replace_char_beam}
\title{Use beam search to replace nucleotides in sequence}
\usage{
replace_char_beam(sequence, model, vocabulary, num_alternatives = 2)
}
\arguments{
\item{sequence}{input sequence}

\item{model}{trained lstm}

\item{vocabulary}{ordered vocabulary of input sequence}

\item{num_alternatives}{number of best solutions}
}
\description{
Use beam search to replace nucleotides in sequence
}
