% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{replaceChar}
\alias{replaceChar}
\title{Replaces specific nucleotides in a sequence}
\usage{
replaceChar(
  sequence,
  model,
  char = "X",
  vocabulary = c("l", "a", "c", "g", "t")
)
}
\arguments{
\item{sequence}{input sequence, length should be in sync with the model.
If length exceeds input.shape of model then only the right side of the
sequence will be used.}

\item{model}{trained model from the function \code{trainNetwork()}}

\item{char}{character in the sequence that will be replaced}

\item{vocabulary}{ordered vocabulary of input sequence}
}
\description{
Replaces specific nucleotides in a sequence
}
\examples{
\dontrun{
example.model <- keras::load_model_hdf5("example_model.hdf5")
replaceChar(sequence = sequence, model = example.model)}
}
