% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{reset_states_cb}
\alias{reset_states_cb}
\title{Reset states callback}
\usage{
reset_states_cb(path_file_log, path_file_logVal)
}
\arguments{
\item{path_file_log}{Path to log of training files.}

\item{path_file_logVal}{Path to log of validation files.}
}
\value{
A keras callback that resets states of LSTM layers.
}
\description{
Reset states at start/end of validation and whenever file changes. Can be used for stateful LSTM.
}
\examples{
\dontshow{if (reticulate::py_module_available("tensorflow")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rs <- reset_states_cb(path_file_log = tempfile(), path_file_logVal = tempfile())
\dontshow{\}) # examplesIf}
}
