% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{reshape_input}
\alias{reshape_input}
\title{Replace input layer}
\usage{
reshape_input(model, input_shape)
}
\arguments{
\item{model}{A keras model.}

\item{input_shape}{The new input shape vector (without batch size).}
}
\description{
Replace first layer of model with new input layer of different shape. Only works for sequential models.
}
