% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{reshape_tensor}
\alias{reshape_tensor}
\title{Reshape tensors for set learning}
\usage{
reshape_tensor(
  x,
  y,
  new_batch_size,
  samples_per_target,
  batch_size,
  path,
  voc_len,
  buffer_len = NULL,
  maxlen,
  reshape_mode = "time_dist",
  concat_maxlen = NULL
)
}
\arguments{
\item{new_batch_size}{Size of first axis of input/targets after reshaping.}

\item{samples_per_target}{How many samples to use for one target.}

\item{batch_size}{Number of samples in one batch.}

\item{path}{Path to input files.}

\item{voc_len}{Length of vocabulary.}

\item{buffer_len}{Only applies if \code{reshape_mode = "concat"}. If buffer_len is an integer, the subsequences are interspaced with buffer_len rows.
The input sequence length is (maxlen \* samples_per_target) + buffer_len \* (samples_per_target - 1).}

\item{maxlen}{Length of predictor sequence.}

\item{reshape_mode}{"time_dist", "multi_input" or "concat".  If reshape_mode is "multi_input",  will produce samples_per_target separate inputs,
each of length maxlen. If reshape_mode is "time_dist", will produce a 4D input array. The dimensions correspond to
(batch_size, samples_per_target, maxlen, length(vocabulary)). If reshape mode is "concat", will concatenate samples_per_target sequences
of length maxlen to one long sequence}
}
\description{
Reshape input x and target y. Aggregates multiple samples from x and y into single input/target batches
}
