% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{savechecks}
\alias{savechecks}
\title{Checkpoints saving function}
\usage{
savechecks(cp, runname, model, optimizer, history)
}
\arguments{
\item{cp}{Type of the checkpoint.}

\item{runname}{Name of the run. Name will be used to identify output from callbacks.}

\item{model}{A keras model.}

\item{optimizer}{A keras optimizer.}

\item{history}{A keras history object.}
}
\description{
Checkpoints saving function
}
\keyword{internal}
