% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{sequenceToArrayLabel}
\alias{sequenceToArrayLabel}
\title{One-hot-encodes integer}
\usage{
sequenceToArrayLabel(
  sequence,
  maxlen,
  vocabulary,
  startInd,
  ambiguous_nuc = "zero",
  nuc_dist = NULL,
  use_quality = FALSE,
  quality_vector = NULL,
  use_coverage = FALSE,
  cov_vector = NULL
)
}
\arguments{
\item{sequence}{Sequence of integers.}

\item{maxlen}{Length of predictor sequence.}

\item{vocabulary}{Set of characters to encode.}

\item{startInd}{Start positions of samples in \code{sequence}.}

\item{ambiguous_nuc}{How to handle nucleotides outside vocabulary, either "zero", "discard" or "equal". If "zero", input gets encoded as zero vector;
if "equal" input is 1/length(vocabulary) x length(vocabulary). If "discard" samples containing nucleotides outside vocabulary get discarded.}

\item{nuc_dist}{Nucleotide distribution.}

\item{use_quality}{Use quality scores.}

\item{quality_vector}{Vector of quality probabilities.}
}
\description{

}
