% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{serialize_genomes}
\alias{serialize_genomes}
\title{preprocesses and serializes genomes located in a folder, exports as hdf5}
\usage{
serialize_genomes(directory, out, format = "fa")
}
\arguments{
\item{directory}{input directory where .fasta files are located}

\item{out}{output directory wher hdf5 files will be written to}

\item{format}{either .txt or .fasta}
}
\description{
preprocesses and serializes genomes located in a folder, exports as hdf5
}
