% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sgdr}
\alias{sgdr}
\title{Stochastic Gradient Descent with Warm Restarts}
\usage{
sgdr(lrmin = 5e-10, lrmax = 0.05, restart = 50, mult = 1, epoch = NULL)
}
\arguments{
\item{lrmin}{Lower limit of the range for the learning rate.}

\item{lrmax}{Upper limit of the range for the learning rate.}

\item{restart}{Number of epochs until a restart is conducted.}

\item{mult}{Factor, by which the number of epochs until a restart is increased at every restart.}

\item{epoch}{Epoch, for which the learning rate shall be calculated.}
}
\value{
A numeric value.
}
\description{
Compute the learning Rate for a given epoch using Stochastic Gradient Descent with Warm Restarts. Implements approach from this \href{https://arxiv.org/abs/1608.03983}{paper}.
}
\examples{
sgdr(lrmin = 5e-10, lrmax = 5e-2, restart = 50,
mult = 1, epoch = 5)

}
