% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_if_no_keras}
\alias{skip_if_no_keras}
\title{Tests if Keras is available for the unit tests}
\usage{
skip_if_no_keras(version = NULL)
}
\arguments{
\item{version}{required keras version}
}
\description{
Tests if Keras is available for the unit tests
}
