% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{splitSequence}
\alias{splitSequence}
\title{Helper function for fastaFileGenerator}
\usage{
splitSequence(seq, vocabulary, maxlen)
}
\arguments{
\item{seq}{a character sequence}

\item{vocabulary}{vector of allowed characters}

\item{maxlen}{sequences in the output < (maxlen + 1) get discarded}
}
\description{
splits a character sequences into vector of sequences
}
