% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu-management.R
\name{start_gpu_session}
\alias{start_gpu_session}
\title{Set # of GPUs to be used in TensorFlow backend.}
\usage{
start_gpu_session(gpus = "0")
}
\arguments{
\item{gpus}{GPU ids}
}
\description{
Set # of GPUs to be used in TensorFlow backend.
}
