% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{start_ind_ignore_amb_single_seq}
\alias{start_ind_ignore_amb_single_seq}
\title{Helper function for get_start_ind, extracts the start positions of all potential samples (considering step size and vocabulary)}
\usage{
start_ind_ignore_amb_single_seq(
  seq,
  maxlen,
  step,
  vocabulary,
  train_mode = "lm"
)
}
\arguments{
\item{seq}{Sequences.}

\item{maxlen}{Length of one sample.}

\item{step}{How often to take a sample.}

\item{vocabulary}{Vector of allowed characters in samples.}

\item{train_mode}{"lm" or "label".}

\item{length_vector}{Length of sequences in \code{seq_vector}.}
}
\description{
Helper function for get_start_ind, extracts the start positions of all potential samples (considering step size and vocabulary)
}
