% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{statesWrapper}
\alias{statesWrapper}
\title{Inference on set of fasta files}
\usage{
statesWrapper(
  function_name = "writeStates",
  function_args = list(),
  fasta.path,
  output_path
)
}
\arguments{
\item{function_name}{Name of base function, either "writeStates", "writeStatesByFastaEntries" or "statesByFastaOneFile".}

\item{function_args}{List of arguments for function in \code{function_name}. If no new argument is given, default values will be used.}

\item{fasta.path}{Path to folder containing fasta files.}

\item{output_path}{Path to folder of output files.}
}
\description{

}
\examples{
\dontrun{
statesWrapper(function_name = "writeStates", function_args = list(model.path = "/path/to/model.h5", padding = TRUE, layer_depth = 1), 
fasta.path = "/path/to/fasta/folder", output_path = "/where/to/write/output")
}
}
