% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{stepdecay}
\alias{stepdecay}
\title{Step Decay}
\usage{
stepdecay(lrmax = 0.005, newstep = 50, mult = 0.7, epoch = NULL)
}
\arguments{
\item{lrmax}{Upper limit of the range for the learning rate.}

\item{newstep}{Number of epochs until the learning rate is reduced.}

\item{mult}{Factor, by which the number of epochs until a restart is decreased after a new step.}

\item{epoch}{Epoch, for which the learning rate shall be calculated.}
}
\description{
Compute the learning Rate for a given epoch using Step Decay.
}
\examples{
stepdecay(lrmax = 0.005, newstep = 50,
mult = 0.7, epoch = 3)

}
