% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{stridecalc}
\alias{stridecalc}
\title{Stride length calculation}
\usage{
stridecalc(maxlen, plen)
}
\arguments{
\item{maxlen}{Length of the input sequence.}

\item{plen}{Length of a patch.}
}
\description{
Compute the optimal length for Stride.
}
\keyword{internal}
