% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{TB_loss_acc}
\alias{TB_loss_acc}
\title{Tensorboard Writer}
\usage{
TB_loss_acc(writer, loss, acc, epoch)
}
\arguments{
\item{writer}{Name of the tensorboard writer function.}

\item{loss}{Computed loss for a given epoch.}

\item{acc}{Computed accracy for a given epoch.}

\item{epoch}{Epoch, for which the values shall be written to the tensorboard.}
}
\description{
Writes the loss and the accuracy for a given epoch to the tensorboard.
}
\keyword{internal}
