% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{tensorboard_cb}
\alias{tensorboard_cb}
\title{tensorboard callback}
\usage{
tensorboard_cb(tensorboard.log, run.name)
}
\arguments{
\item{tensorboard.log}{Path to tensorboard log directory.}

\item{run.name}{Name of the run (without file ending). Name will be used to identify output from callbacks.}
}
\description{
tensorboard callback
}
