% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{tsneFromStates}
\alias{tsneFromStates}
\title{run the tSNE on a cell from the states.list, return list which contains the output of the Rtsne and the sample ids}
\usage{
tsneFromStates(states.list = NULL,
  states.folder = "example_files/states", cell.number = 2,
  flatten = FALSE, perplexity = 1, pca = TRUE, max_iter = 1000)
}
\arguments{
\item{states.list}{list of states generated by generateStatesromFolder}

\item{states.folder}{generat states list from folder containing tate files}

\item{cell.number}{number of cell/neuron that should go into tSNE}

\item{flatten}{if true, all cells will be used}
}
\description{
run the tSNE on a cell from the states.list, return list which contains the output of the Rtsne and the sample ids
}
