% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{validation_after_training_cb}
\alias{validation_after_training_cb}
\title{Do validation only after training finished}
\usage{
validation_after_training_cb(gen.val, validation_steps)
}
\description{
Do validation only after training finished
}
