% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_viewer.R
\name{visualizePrediction}
\alias{visualizePrediction}
\title{This is a Shiny web application for the visualization of hidden states of
LSTM models from the package deepG. It also comes with 
the possibility to use own trained models.}
\usage{
visualizePrediction(
  sample = "",
  states = FALSE,
  model.path = "",
  fasta.path = "",
  states.path = "",
  start_position = 0,
  end_position = 800,
  batch.size = 200,
  vocabulary = c("l", "a", "g", "c", "t"),
  layer_depth = 1,
  cell_number = 1,
  step = 1,
  padding = TRUE
)
}
\arguments{
\item{sample}{character input string of text}

\item{states}{if states = TRUE, then load states from states_path}

\item{model.path}{path to trained model}

\item{fasta.path}{path to fasta files}

\item{states.path}{path to the .h5 file with the calculated states}

\item{start_position}{start from the dygraph}

\item{end_position}{end from the dygraph}

\item{batch.size}{number of samples to evaluate at once}

\item{vocabulary}{used vocabulary}

\item{layer_depth}{depth of layer to evaluate}

\item{cell_number}{cell_number showed in the dygraph}

\item{step}{frequency of sampling steps}

\item{padding}{TRUE/FALSE generate states for the first maxlen nucleotides}
}
\description{
This is a Shiny web application for the visualization of hidden states of
LSTM models from the package deepG. It also comes with 
the possibility to use own trained models.
}
