% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeDict}
\alias{writeDict}
\title{Generates dictionary for LSTMVis}
\usage{
writeDict(dat, filename = "train.dict")
}
\arguments{
\item{dat}{filename where dict file is written to}

\item{filename}{file name}
}
\description{
Generates dictionary for LSTMVis
}
