% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeHdf5}
\alias{writeHdf5}
\title{generates hdf5 file containing the character id for each time step}
\usage{
writeHdf5(dat, filename = "train.hdf5", verbose = F)
}
\arguments{
\item{dat}{character}

\item{filename}{filename where hdf5 file is written to}

\item{verbose}{TRUE/FALSE}
}
\description{
generates hdf5 file containing the character id for each time step
}
