% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{writeStatesByFastaEntries}
\alias{writeStatesByFastaEntries}
\title{Write states to h5 file}
\usage{
writeStatesByFastaEntries(
  model.path = NULL,
  layer.depth = NULL,
  layer_name = NULL,
  fasta.path,
  round_digits = 2,
  file_name = "states.h5",
  file_path,
  step = 1,
  vocabulary = c("a", "c", "g", "t"),
  batch.size = 256,
  padding = FALSE,
  file_type = "h5",
  model = NULL,
  mode = "lm",
  target_middle = FALSE,
  format = "fasta"
)
}
\arguments{
\item{model.path}{Path to a pretrained model.}

\item{layer.depth}{Depth of layer to evaluate. If NULL last layer is used.}

\item{layer_name}{Name of layer to get output from.}

\item{fasta.path}{Path to fasta file.}

\item{round_digits}{Number of decimal places.}

\item{file_name}{Filename to store states, function adds "Nr" + "i" before name, where i is entry number.}

\item{file_path}{Path to folder, where to write output.}

\item{step}{Frequency of sampling steps.}

\item{vocabulary}{Vector of allowed characters, character outside vocabulary get encoded as 0-vector.}

\item{batch.size}{Number of samples to evaluate at once. Does not change output, only relevant for speed and memory.}

\item{padding}{Logical scalar, generate states for first maxlen nucleotides by
padding beginning of sequence with 0-vectors.}

\item{file_type}{Either "h5" or "csv".}

\item{model}{A keras model. If model and model.path are not NULL, model will be used for inference.}

\item{mode}{Either "lm" for language model or "label" for label classification.}

\item{target_middle}{Whether target to predict is in middle of sequence.}

\item{format}{Either "fasta" or "fastq".}
}
\description{
\code{writeStatesByFastaEntries} Removes layers (optional) from pretrained model and calculates states of fasta file, writes a separate 
h5 file for every fasta entry in fasta file. h5 files also contain the nucleotide sequence and positions of targets corresponding to states.      
Names of output files are: file_path + "Nr" + i + filename + file_type, where i is the number of the fasta entry.
}
