% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{calculateStepsPerEpoch}
\alias{calculateStepsPerEpoch}
\title{do one full preprocessing iteration to figure out what the observed
steps_per_epoch value is}
\usage{
calculateStepsPerEpoch(dir, batch.size = 256, maxlen = 250,
  format = "fasta")
}
\description{
do one full preprocessing iteration to figure out what the observed
steps_per_epoch value is
}
