% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{checkpoint_cb}
\alias{checkpoint_cb}
\title{Checkpoint callback}
\usage{
checkpoint_cb(filepath_checkpoints, save_weights_only, save_best_only)
}
\arguments{
\item{save_weights_only}{Whether to save weights only.}

\item{save_best_only}{Only save model that improved on best val_loss score.}
}
\description{
Checkpoint callback
}
