% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{conf_matrix_cb}
\alias{conf_matrix_cb}
\title{Confusion matrix callback.}
\usage{
conf_matrix_cb(path_tensorboard, run_name, confMatLabels, cm_dir, total_epochs)
}
\arguments{
\item{path_tensorboard}{Path to tensorboard directory}

\item{run_name}{Name of training run.}

\item{confMatLabels}{Names of classes.}

\item{cm_dir}{Directory that contains confusion matrix files.}

\item{total_epochs}{Number of total training epochs.}
}
\description{
Create a confusion matrix to display in tensorboard "IMAGES" tab.
}
