% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model_wavenet}
\alias{create_model_wavenet}
\title{Create wavenet model}
\usage{
create_model_wavenet(
  filters = 16,
  kernel_size = 2,
  residual_blocks,
  maxlen,
  input_tensor = NULL,
  initial_kernel_size = 32,
  initial_filters = 32,
  output_channels = 4,
  output_activation = "softmax",
  solver = "adam",
  learning_rate = 0.001,
  compile = TRUE,
  verbose = TRUE,
  model_seed = NULL,
  mixed_precision = FALSE,
  mirrored_strategy = NULL
)
}
\arguments{
\item{filters}{Number of filters. For multiple layers, assign a vector.}

\item{kernel_size}{Size of 1d convolutional layers. For multiple layers, assign a vector. (e.g, \code{rep(3,2)} for two layers and kernel size 3)}

\item{maxlen}{Length of predictor sequence.}

\item{solver}{Optimization method, options are \verb{"adam", "adagrad", "rmsprop"} or \code{"sgd"}.}

\item{learning_rate}{Learning rate for optimizer.}

\item{compile}{Whether to compile the model.}

\item{verbose}{Boolean.}

\item{model_seed}{Set seed for model parameters in tensorflow if not \code{NULL}.}

\item{mixed_precision}{Whether to use mixed precision (https://www.tensorflow.org/guide/mixed_precision).}

\item{mirrored_strategy}{Whether to use distributed mirrored strategy. If NULL, will use distributed mirrored strategy only if >1 GPU available.}
}
\description{
Creates network architecture as described \href{https://arxiv.org/abs/1609.03499}{here}. Implementation
uses code from \href{https://github.com/r-tensorflow/wavenet}{here}.
}
\examples{
\dontrun{
model <- create_model_wavenet(residual_blocks = 2^rep(1:4, 2), maxlen = 1000)
}
}
