% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model_wavenet}
\alias{create_model_wavenet}
\title{Create wavenet model}
\usage{
create_model_wavenet(
  filters = 16,
  kernel_size = 2,
  residual_blocks,
  maxlen,
  input_tensor = NULL,
  initial_kernel_size = 32,
  initial_filters = 32,
  output_channels = 4,
  output_activation = "softmax",
  solver = "adam",
  learning_rate = 0.001,
  compile = TRUE,
  verbose = TRUE,
  model_seed = NULL
)
}
\arguments{
\item{filters}{Integer, the dimensionality of the output space (i.e. the
number of output filters in the convolution).}

\item{kernel_size}{An integer or list of a single integer, specifying the
length of the 1D convolution window.}

\item{residual_blocks}{if a single integer: the number of residual blocks in the
model - the dilation rate of each block `i` is calculated by `2^i`. If a vector,
then it's used as the dilation rate of each residual block.}

\item{maxlen}{Length of predictor sequence.}

\item{input_tensor}{Tensor to use as input for the model. Usually a 3d tensor.}

\item{initial_kernel_size}{kernel size of the first causal convolution.}

\item{initial_filters}{number of filters in the initial causal convolution.}

\item{output_channels}{number of channels in the output.}

\item{output_activation}{activation function for the last layer. Default: 'softmax'}

\item{solver}{Optimization method, options are \verb{"adam", "adagrad", "rmsprop"} or \code{"sgd"}.}

\item{learning_rate}{Learning rate for optimizer.}

\item{compile}{Whether to compile the model.}

\item{verbose}{Boolean.}

\item{model_seed}{Set seed for model parameters in tensorflow if not \code{NULL}.}
}
\description{
Creates network architecture as described \href{https://arxiv.org/abs/1609.03499}{here}. Implementation
uses code from \href{https://github.com/r-tensorflow/wavenet}{here}.
}
\examples{
\dontrun{
model <- create_model_wavenet(residual_blocks = 2^rep(1:4, 2), maxlen = 1000)
}
}
