% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crispr_full}
\alias{crispr_full}
\title{CRISPR data subset}
\format{
Large character of 7.50 MB
}
\usage{
data(crispr_full)
}
\description{
Example training dataset consisting of a sequence of nucleotides of CRISPR loci.
Filtered for unambiguous characters and contains only characters in the vocabulary \{A,G,G,T
\}.
contain all CRISPR loci found in NCBI representative genomes using CRT.
Can be loaded to workspace via \code{data(crispr_full)}.
}
\references{
\url{https://github.com/philippmuench}
}
\keyword{datasets}
