% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crispr_sample}
\alias{crispr_sample}
\title{CRISPR data}
\format{
Large character of 442.41 kB
}
\usage{
crispr_sample
}
\description{
Example training dataset consisting of a sequence of nucleotides of CRISPR loci
Filtered for unambiguous characters and contains only characters in the vocabulary {A,G,G,T
}
Can be loaded to workspace via data(crispr_sample)
}
\references{
\url{http://github.com/philippmuench}
}
\keyword{datasets}
