% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{download_from_ncbi}
\alias{download_from_ncbi}
\title{Download genome from NCBI and returns preprocessed object}
\usage{
download_from_ncbi(accession = "GCF_000007825.1", ...)
}
\arguments{
\item{accession}{RefSeq assembly accession e.g. 'GCA_000196515.1' forhttps://www.ncbi.nlm.nih.gov/assembly/GCA_000196515.1 . A collection of RefSeq accessions can be generated
using https://www.ncbi.nlm.nih.gov/genome/browse}
}
\description{
Download genome from NCBI and returns preprocessed object
}
