% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{early_stopping_time_cb}
\alias{early_stopping_time_cb}
\title{Stop training callback}
\usage{
early_stopping_time_cb(stop_time = NULL)
}
\arguments{
\item{stop_time}{Time in seconds after which to stop training.}
}
\description{
Stop training after specified time.
}
