% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli}
\alias{ecoli}
\title{E. coli genome for evaluation
Can be loaded to workspace via data(ecoli)}
\format{
Large character of 4.4 Mb
}
\usage{
ecoli
}
\description{
E. coli genome for evaluation
Can be loaded to workspace via data(ecoli)
}
\references{
\url{https://science.sciencemag.org/content/277/5331/1453.long}
}
\keyword{datasets}
