% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_small}
\alias{ecoli_small}
\title{subset of the E. coli genome for evaluation
Can be loaded to workspace via data(ecoli_small)}
\format{
character
}
\usage{
ecoli_small
}
\description{
subset of the E. coli genome for evaluation
Can be loaded to workspace via data(ecoli_small)
}
\references{
\url{https://science.sciencemag.org/content/277/5331/1453.long}
}
\keyword{datasets}
