% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorboard.R
\name{ecoliEvaluation}
\alias{ecoliEvaluation}
\title{Runs a evaluation on the E. coli genome, returns a list of accuracy value for each letter in the vocabulary}
\usage{
ecoliEvaluation(model, genome = "ecoli_small", maxlen = 80,
  vocabulary = c("-", "|", "a", "c", "g", "t"))
}
\arguments{
\item{maxlen}{Time steps to unroll for (e.g. length of semi-redundant chunks)}

\item{vocabulary}{Char contains the vocabulary from the input char, it should be sorted.}
}
\description{
Runs a evaluation on the E. coli genome, returns a list of accuracy value for each letter in the vocabulary
}
