% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluates a trained model on fasta/fastq or rds files}
\usage{
evaluate_model(
  path_input,
  model = NULL,
  batch_size = 100,
  step = 1,
  padding = FALSE,
  vocabulary = c("a", "c", "g", "t"),
  vocabulary_label = list(c("a", "c", "g", "t")),
  number_batches = 10,
  format = "fasta",
  target_middle = FALSE,
  mode = "lm",
  output_format = "target_right",
  ambiguous_nuc = "zero",
  evaluate_all_files = FALSE,
  verbose = TRUE,
  max_iter = 20000,
  target_from_csv = NULL,
  max_samples = NULL,
  proportion_per_seq = NULL,
  seed = 1234,
  auc = FALSE,
  auprc = FALSE,
  exact_num_samples = NULL,
  ...
)
}
\arguments{
\item{path_input}{Input directory where fasta/fastq or rds files are located.}

\item{model}{A keras model.}

\item{batch_size}{Number of samples per batch.}

\item{step}{How often to take a sample.}

\item{padding}{Whether to pad sequences too short for one sample with zeros.}

\item{vocabulary}{Vector of allowed characters, character outside vocabulary get encoded as specified in ambiguous_nuc.}

\item{vocabulary_label}{List of labels for targets of each output layer.}

\item{number_batches}{How many batches to evaluate.}

\item{format}{File format, "fasta", "fastq" or "rds".}

\item{mode}{Either "lm" for language model and "label_header", "label_csv" or "label_folder" for label classification.}

\item{output_format}{Determines shape of output tensor for language model.
Either "target_right", "target_middle_lstm", "target_middle_cnn" or "wavenet".
Assume a sequence "AACCGTA". Output correspond as follows
"target_right": X = "AACCGT", Y = "A"
"target_middle_lstm": X = (X_1 = "AAC", X_2 = "ATG"), Y = "C" (note reversed order of X_2)
"target_middle_cnn": X = "AACGTA", Y = "C"
"wavenet": X = "AACCGT", Y = "ACCGTA"}

\item{ambiguous_nuc}{How to handle nucleotides outside vocabulary, either "zero", "discard", "empirical" or "equal". If "zero", input gets encoded as zero vector;
if "equal" input is 1/length(vocabulary) x length(vocabulary). If "discard" samples containing nucleotides outside vocabulary get discarded.
If "empirical" use nucleotide distribution of current file.}

\item{evaluate_all_files}{Boolean, if TRUE will iterate over all files in \code{path_input} once. \code{number_batches} will be overwritten.}

\item{verbose}{Whether to show message.}

\item{max_iter}{Stop after max_iter number of iterations failed to produce a new batch.}

\item{target_from_csv}{Path to csv file with target mapping. One column should be called "file" and other entries in row are the targets.}

\item{max_samples}{Maximum number of samples to use from one file. If not NULL and file has more than \code{max_samples} samples, will randomly choose a
subset of \code{max_samples} samples.}

\item{proportion_per_seq}{Numerical value between 0 and 1. Proportion of possible samples to take from one file. Takes samples from random subsequence.}

\item{seed}{Sets seed for set.seed function, for reproducible results when using \code{shuffle_file_order} or \code{shuffle_input}}

\item{auc}{Whether to include AUC metric. Only possible for 2 targets if layer activation is "softmax".}

\item{auprc}{Whether to include AUPRC metric. Only possible for 2 targets if layer activation is "softmax".}

\item{exact_num_samples}{Exact number of samples to evaluate. If you want to evaluate a number of samples not devisible by batch_size. Useful if you want
to evaluate a data set exactly ones and know the number of samples already. Should be a vector if mode = "label_folder" (with same length as vocabulary_label)
and else an integer.}

\item{...}{Further generator options.}
}
\description{
Returns evaluation metric like confusion matrix, loss, AUC, AUPRC, MAE, MSE (depending and output layer).
Evaluates \code{batch_size} * \code{number_batches} samples.
}
