% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{fasta_files_generator}
\alias{fasta_files_generator}
\title{custom generator for fasta files}
\usage{
fasta_files_generator(dir, format = "fasta")
}
\arguments{
\item{directory}{input directory where .fasta files are located}
}
\description{
custom generator for fasta files
}
