% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{fastaFileGenerator}
\alias{fastaFileGenerator}
\title{Custom generator for fasta files, will produce chunks in size of batch.size
by iterating over the input files.}
\usage{
fastaFileGenerator(corpus.dir, format = "fasta", batch.size = 256,
  maxlen = 250, max_iter = 20, seqStart = "l", seqEnd = "l",
  withinFile = "p", vocabulary = c("l", "p", "a", "c", "g", "t"),
  verbose = FALSE, randomFiles = FALSE, step = 1,
  showWarnings = FALSE)
}
\arguments{
\item{corpus.dir}{input directory where .fasta files are located}

\item{format}{file format}

\item{batch.size}{number of samples}

\item{maxlen}{length of one sample}

\item{max_iter}{stop after max_iter number of iterations failed to produce new sample}

\item{verbose}{TRUE/FALSE show information about files and running time}

}
\description{
Custom generator for fasta files, will produce chunks in size of batch.size
by iterating over the input files.
}
