% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{fastaFileGenerator}
\alias{fastaFileGenerator}
\title{custom generator for fasta files, will produce chunks in size of batch.size
by iterating over the input files.}
\usage{
fastaFileGenerator(corpus.dir, format = "fasta", batch.size = 512,
  maxlen = 250, max_iter = 20, vocabulary = c("-", "|", "a", "c",
  "g", "t"), verbose = FALSE)
}
\arguments{
\item{corpus.dir}{input directory where .fasta files are located}

\item{format}{file format}

\item{batch.size}{number of samples}

\item{maxlen}{length of one sample}

\item{max_iter}{stop after max_iter number of iterations failed to produce new sample}

\item{verbose}{TRUE/FALSE show information about files and running time}
}
\description{
custom generator for fasta files, will produce chunks in size of batch.size
by iterating over the input files.
}
