% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{fastaFileGenerator}
\alias{fastaFileGenerator}
\title{custom generator for fasta files, will produce chunks in size of batch.size
by iterating over the input files. If the input file is smaller than the
batch.size, batch size will be reduced to the maximal size. So the last batch
of a file is usually smaller.}
\usage{
fastaFileGenerator(dir, format = "fasta", batch.size = 512,
  maxlen = 80, verbose = F)
}
\arguments{
\item{directory}{input directory where .fasta files are located}
}
\description{
see https://github.com/bagasbgy/kerasgenerator/blob/master/R/timeseries_generator.R
}
