% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{fastaFileGenerator}
\alias{fastaFileGenerator}
\title{Custom generator for fasta files, will produce chunks in size of batch.size
by iterating over the input files.}
\usage{
fastaFileGenerator(corpus.dir, format = "fasta", batch.size = 256,
  maxlen = 250, max_iter = 20, seqStart = "l", seqEnd = "l",
  withinFile = "p", vocabulary = c("l", "p", "a", "c", "g", "t"),
  verbose = FALSE, randomFiles = FALSE, step = 1,
  showWarnings = FALSE, seed = 1234, shuffleFastaEntries = FALSE)
}
\arguments{
\item{corpus.dir}{input directory where .fasta files are located}

\item{format}{file format}

\item{batch.size}{number of samples}

\item{maxlen}{length of one sample}

\item{max_iter}{stop after max_iter number of iterations failed to produce new sample}

\item{seqStart}{insert character at beginning of sequence}

\item{seqEnd}{insert character at end of sequence}

\item{withinFile}{insert characters within sequence}

\item{vocabulary}{vector of allowed characters, samples with other chars get discarded}

\item{randomFiles}{TRUE/FALSE, whether to go through files randomly or sequential}

\item{step}{how often to take a sample}

\item{showWarnings}{TRUE/FALSE, give warning if character outside vocabulary appears}

\item{seed}{argument for set.seed function, for reproducible results}

\item{shuffleFastaEntries}{shuffle entries in every fasta file}
}
\description{
Custom generator for fasta files, will produce chunks in size of batch.size
by iterating over the input files.
}
