% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{focal_loss_multiclass}
\alias{focal_loss_multiclass}
\title{Focal loss for two or more labels}
\usage{
focal_loss_multiclass(y_true, y_pred, gamma = 2.5, alpha = c(1))
}
\arguments{
\item{y_true}{Vector of true values.}

\item{y_pred}{Vector of predicted values.}

\item{gamma}{Focusing parameter.}

\item{alpha}{Vector of weighting factors.}
}
\description{
Focal loss for two or more labels
}
\examples{
y_true <- matrix(c(0, 1, 0, 0, 0, 1), nrow = 2, byrow = TRUE)
y_pred <- matrix(c(0.15, 0.8, 0.05,
                   0.08, 0.02, 0.9), nrow = 2, byrow = TRUE) 
fl <- focal_loss_multiclass(y_true, y_pred)
fl$numpy()

}
