% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{function_args_cb}
\alias{function_args_cb}
\title{Function arguments callback}
\usage{
function_args_cb(argumentList, path_tensorboard, run_name)
}
\arguments{
\item{path_tensorboard}{Path to path_tensorboard directory or NULL. If NULL, training not tracked on tensorboard.}

\item{run_name}{Name of the run. Name will be used to identify output from callbacks. If NULL, will use date as run name. If name already present, will add "_2" to name
or "_{x + 1}" if name ends with integer x.}
}
\description{
Print train_model call in text field of tensorboard.
}
