% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{function_args_cb}
\alias{function_args_cb}
\title{Function arguments callback}
\usage{
function_args_cb(argumentList, path_tensorboard, run_name)
}
\arguments{
\item{argumentList}{List of function arguments.}

\item{path_tensorboard}{Path to tensorboard directory.}

\item{run_name}{Name of training run.}
}
\description{
Print \code{\link{train_model}} call in tensorboard "TEXT" tab.
}
