% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{generateStatesFromFolder}
\alias{generateStatesFromFolder}
\title{generate states list for tSNE from a fasta folder
it is expected that the length of each fasta file is equal}
\usage{
generateStatesFromFolder(fasta.folder = "example_files/fasta",
  model.path = "example_files/dummy_model_cpu.hdf5")
}
\arguments{
\item{fasta.folder}{path to fasta folder}

\item{model.path}{path to model file}

\item{use.names}{save id of fasta header}
}
\description{
generate states list for tSNE from a fasta folder
it is expected that the length of each fasta file is equal
}
