% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generator_fasta_label_folder_wrapper}
\alias{generator_fasta_label_folder_wrapper}
\title{Generator wrapper}
\usage{
generator_fasta_label_folder_wrapper(
  val,
  new_batch_size = NULL,
  batch_size = NULL,
  path = NULL,
  voc_len = NULL,
  maxlen = NULL,
  set_learning = NULL
)
}
\arguments{
\item{val}{Train or validation generator.}

\item{new_batch_size}{Only applied if \code{set_learning} is not NULL}

\item{batch_size}{Number of samples in one batch.}

\item{path}{Path to input files.}

\item{voc_len}{Length of vocabulary.}

\item{maxlen}{Length of predictor sequence.}

\item{set_learning}{When you want to assign one label to set of samples. Only implemented for \code{train_type = "label_folder"}.
Input is a list with the following parameters}
}
\description{
Combines generators created by \code{\link{generator_initialize}} into a single generator.
}
