% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generator_rds}
\alias{generator_rds}
\title{Read batches from rds arrays.}
\usage{
generator_rds(
  rds_folder,
  batch_size,
  path_file_log = NULL,
  max_samples = NULL,
  proportion_per_seq = NULL,
  target_len = NULL,
  sample_by_file_size = FALSE,
  n_gram = NULL,
  n_gram_stride = 1,
  add_noise = NULL
)
}
\arguments{
\item{rds_folder}{Path to input files.}
}
\description{
Read batches from rds arrays.
}
