% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generator_rds}
\alias{generator_rds}
\title{Rds data generator}
\usage{
generator_rds(
  rds_folder,
  batch_size,
  path_file_log = NULL,
  max_samples = NULL,
  proportion_per_seq = NULL,
  target_len = NULL,
  reverse_complement = FALSE,
  sample_by_file_size = FALSE,
  n_gram = NULL,
  n_gram_stride = 1,
  add_noise = NULL
)
}
\arguments{
\item{rds_folder}{Path to input files.}

\item{batch_size}{Number of samples in one batch.}

\item{path_file_log}{Write name of files to csv file if path is specified.}

\item{max_samples}{Maximum number of samples to use from one file. If not \code{NULL} and file has more than \code{max_samples} samples, will randomly choose a
subset of \code{max_samples} samples.}

\item{proportion_per_seq}{Numerical value between 0 and 1. Proportion of sequence to take samples from (use random subsequence).}

\item{reverse_complement}{Boolean, for every new file decide randomly to use original data or its reverse complement.}

\item{sample_by_file_size}{Sample new file weighted by file size (bigger files more likely).}

\item{n_gram}{Integer, encode target not nucleotide wise but combine n nucleotides at once. For example for \verb{n=2, "AA" ->  (1, 0,..., 0),}
\verb{"AC" ->  (0, 1, 0,..., 0), "TT" -> (0,..., 0, 1)}, where the one-hot vectors have length \code{length(vocabulary)^n}.}

\item{n_gram_stride}{Step size for n-gram encoding. For AACCGGTT with \code{n-gram = 4} and \code{n_gram_stride = 2}, generator encodes
\verb{(AACC), (CCGG), (GGTT)}; for \code{n_gram_stride = 4} generator encodes \verb{(AACC), (GGTT)}.}

\item{add_noise}{\code{NULL} or list of arguments. If not \code{NULL}, list must contain the following arguments: \code{noise_type} can be \code{"normal"} or \code{"uniform"};
optional arguments \code{sd} or \code{mean} if noise_type is \code{"normal"} (default is \code{sd=1} and \code{mean=0}) or \verb{min, max} if \code{noise_type} is \code{"uniform"}
(default is \verb{min=0, max=1}).}
}
\description{
Creates training batches from rds files. Rds files must contain a
list of length 2 (input/target) or of length 1 (for language model).
If \code{target_len} is not NULL will take the last \code{target_len} entries of
the first list element as targets and the rest as input.
}
