% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_class_weight}
\alias{get_class_weight}
\title{Estimate frequency of different classes}
\usage{
get_class_weight(
  path,
  labelVocabulary = NULL,
  format = "fasta",
  file_proportion = 1,
  train_type = "label_folder",
  csv_path = NULL
)
}
\arguments{
\item{labelVocabulary}{Character vector of possible targets. Targets outside \code{labelVocabulary} will get discarded.}

\item{format}{File format, either fasta or fastq.}

\item{file_proportion}{Proportion of files to randomly sample for estimating class distributions.}

\item{train_type}{Either "label_folder", "label_header" or "label_csv".}
}
\description{
Estimate frequency of different classes
}
