% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_run_name}
\alias{get_run_name}
\title{Generate run_name if none is given or is already present.}
\usage{
get_run_name(
  run_name = NULL,
  path_tensorboard,
  path_checkpoint,
  path_log,
  path_model_card,
  auto_extend = FALSE
)
}
\arguments{
\item{auto_extend}{If run_name is already present, add "_2" to name. If name already ends with "_x" replace x with x+1.}
}
\description{
If no run name is given, will use date as run name. If run name is already present will add _2 to name or
_x+1 if name ends with _x and x is integer.
}
\keyword{internal}
