% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_start_ind}
\alias{get_start_ind}
\title{Computes start position of samples}
\usage{
get_start_ind(
  seq_vector,
  length_vector,
  maxlen,
  step,
  train_mode = "label",
  discard_amb_nuc = FALSE,
  vocabulary = c("A", "C", "G", "T")
)
}
\arguments{
\item{seq_vector}{Vector of character sequences.}

\item{length_vector}{Length of sequences in \code{seq_vector}.}

\item{maxlen}{Length of one predictor sequence.}

\item{step}{Distance between samples from one entry in \code{seq_vector}.}

\item{train_mode}{Either "lm" for language model or "label" for label classification. Language models need one character more
(the target) for one sample.}

\item{discard_amb_nuc}{Discard all samples that contain characters outside vocabulary.}
}
\description{
Helper function for data generators. 
Computes start positions in sequence where samples can be extracted, given maxlen, step size and ambiguous nucleotide constraints.
}
