% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_vocabulary}
\alias{get_vocabulary}
\title{Get vocabulary from character string}
\usage{
get_vocabulary(char)
}
\arguments{
\item{char}{a character string of text, length of one}
}
\description{
Use this function with a character string as input such as data(train)
}
