% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{getstates}
\alias{getstates}
\title{Get cell state responses of a model to a preprocessed text corpus}
\usage{
getstates(model_path, x, maxlen = 30, batch_size = 100,
  run_name = "output", type = "csv")
}
\arguments{
\item{x}{preprocessed dataset for prediction}

\item{maxlen}{time steps to unroll for (e.g. length of semi-redundant chunks)}

\item{batch_size}{how many chunks are trained in parallel}

\item{run_name}{name of output files without ending}

\item{type}{will save as hdf5 if type is set to 'hdf5', otherwise as csv}

\item{model}{LSTM model in hdf5 format}
}
\description{
Get cell state responses of a model to a preprocessed text corpus
}
