% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-states.R
\name{getstates}
\alias{getstates}
\title{Get cell states of semi-redundant chunks}
\usage{
getstates(model_path, x, maxlen = 30, batch_size = 100,
  run_name = "output", type = "csv")
}
\arguments{
\item{x}{semi-redundant chunks (one-hot)}

\item{maxlen}{time steps to unroll for}

\item{batch_size}{how many samples are trained in parallel}

\item{run_name}{name of output files without ending}

\item{type}{will save as hdf5 if type is set to 'hdf5', otherwise as csv}

\item{model}{keras model in hdf5 format}
}
\description{
Get cell states of semi-redundant chunks
}
