% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-states.R
\name{getStates}
\alias{getStates}
\title{Get cell states of semi-redundant chunks}
\usage{
getStates(model.path, x, maxlen = 30, batch.size = 100,
  run.name = "output", type = "csv", verbose = F)
}
\arguments{
\item{x}{Semi-redundant chunks (one-hot)}

\item{maxlen}{Time steps to unroll for}

\item{type}{Save-type, will save as hdf5 if type is set to 'hdf5' (default .csv)}

\item{verbose}{TRUE/FALSE}

\item{model_path}{Path to keras model in hdf5 format}

\item{batch_size}{How many samples are trained in parallel}

\item{run_name}{Name of output files without ending}
}
\description{
Get cell states of semi-redundant chunks
}
