% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-states.R
\name{getStatesFromFasta}
\alias{getStatesFromFasta}
\title{Get cell states from a fasta file}
\usage{
getStatesFromFasta(model.path = "example_files/dummy_model_cpu.hdf5",
  fasta.path = "example_files/fasta/a.fasta", maxlen = 80,
  batch.size = 100, verbose = T)
}
\arguments{
\item{model.path}{path to keras model in hdf5 format}

\item{fasta.path}{path to fasta file}

\item{maxlen}{time steps to unroll for}

\item{batch.size}{how many subsequences are predicted in parallel}

\item{verbose}{print output}
}
\description{
Get cell states from a fasta file
}
