% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{getVocabulary}
\alias{getVocabulary}
\title{Returns the vocabulary from character string}
\usage{
getVocabulary(char, verbose = F)
}
\arguments{
\item{char}{Character string of text with the length of one}

\item{verbose}{TRUE/FALSE}
}
\description{
Use this function with a character string.
}
