% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{labelByFolderGeneratorWrapper}
\alias{labelByFolderGeneratorWrapper}
\title{Generator wrapper}
\usage{
labelByFolderGeneratorWrapper(
  val,
  new_batch_size = NULL,
  samples_per_target = NULL,
  batch.size = NULL,
  path = NULL,
  voc_len = NULL,
  maxlen = NULL,
  reshape_mode = NULL
)
}
\arguments{
\item{val}{Train or validation generator.}

\item{path}{Path}
}
\description{

}
