% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R, R/custom_layers.R
\name{layer_aggregate_time_dist_wrapper}
\alias{layer_aggregate_time_dist_wrapper}
\title{Add layer}
\usage{
layer_aggregate_time_dist_wrapper(
  load_r6 = FALSE,
  method = "sum",
  multi_in = FALSE
)

layer_aggregate_time_dist_wrapper(
  load_r6 = FALSE,
  method = "sum",
  multi_in = FALSE
)
}
\arguments{
\item{load_r6}{Whether to load the R6 layer class.}

\item{method}{At least one of the options, \verb{"sum", "max"} or \code{"mean"}.}

\item{multi_in}{Whether to aggregate for a model with multiple inputs (and shared weights).}
}
\value{
A keras layer applying pooling operation(s).
}
\description{
Add output of time distribution representations.

Aggregate output of time distribution representations using sum, max and/or mean function.
}
\examples{
\dontshow{if (reticulate::py_module_available("tensorflow")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
l <- layer_aggregate_time_dist_wrapper() 
\dontshow{\}) # examplesIf}
}
