% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{load_prediction}
\alias{load_prediction}
\title{Read states from h5 file}
\usage{
load_prediction(
  h5_path,
  rows = NULL,
  verbose = TRUE,
  get_target_positions = FALSE,
  get_seq = FALSE
)
}
\arguments{
\item{h5_path}{Path to h5 file.}

\item{rows}{Range of rows to read.  If NULL read everything.}

\item{get_target_positions}{Return position of corresponding targets if TRUE.}

\item{get_seq}{Return nucleotide sequence if TRUE.}
}
\description{
Reads rows from h5 file created by  \code{\link{predict_model}} function,
rows correspond to position in sequence and columns to neurons.
}
