% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{merge_models}
\alias{merge_models}
\title{Merge two models}
\usage{
merge_models(
  models,
  layer_names,
  layer_dense,
  solver = "adam",
  learning.rate = 1e-04,
  freeze_base_model = c(FALSE, FALSE)
)
}
\arguments{
\item{models}{List of two models}

\item{layer_names}{Vector of length 2 with names of layers to merge.}
}
\description{
Merge two models
}
