% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{n_gram_of_matrix}
\alias{n_gram_of_matrix}
\title{One-hot encoding matrix to n-gram encoding matrix}
\usage{
n_gram_of_matrix(input_matrix, n = 3)
}
\arguments{
\item{input_matrix}{Matrix with one 1 per row and zeros otherwise.}

\item{n}{Length of one n-gram.}
}
\description{
One-hot encoding matrix to n-gram encoding matrix
}
\examples{
\dontrun{
x <- c(0,0,1,3,3) 
input_matrix <- keras::to_categorical(x, 4)
n_gram_of_matrix(input_matrix, n = 2) 
}
}
