% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parenthesis}
\alias{parenthesis}
\title{Training dataset of synthetic parenthesis language
Can be loaded to workspace via data(parenthesis)}
\format{
Large character of 4.9 Mb
}
\usage{
parenthesis
}
\description{
Training dataset of synthetic parenthesis language
Can be loaded to workspace via data(parenthesis)
}
\references{
\url{http://github.com/philippmuench}
}
\keyword{datasets}
