% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parenthesis}
\alias{parenthesis}
\title{Parenthesis data}
\format{
Large character of 1.00 MB
}
\usage{
parenthesis
}
\description{
Training dataset of synthetic parenthesis language.
Can be loaded to workspace via \code{data(parenthesis)}.
}
\references{
\url{http://github.com/philippmuench}
}
\keyword{datasets}
