% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{plotTsne}
\alias{plotTsne}
\title{generate and plot tSNE}
\usage{
plotTsne(states, perplexity = c(0.5, 1), pca = TRUE, max_iter = 1000,
  path.metadata = NULL, sample.name.column = 1,
  label.name.column = 2, color.column = 3, out.path = "tsne.pdf")
}
\arguments{
\item{perplexity}{Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation)}

\item{pca}{Whether an initial PCA step should be performed (default: TRUE)}

\item{max_iter}{Number of iterations (default: 1000)}

\item{path.metadata}{file used for coloring}

\item{sample.name.column}{colun in metadata file that holds the sample name}

\item{label.name.column}{column that holds the label name in path.metadata}

\item{color.column}{column that holds the color information in path.metadata}

\item{out.path}{path to output pdf}
}
\description{
generate and plot tSNE
}
