% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictNextNucleotide}
\alias{predictNextNucleotide}
\title{Predict next nucleotides in sequence}
\usage{
predictNextNucleotide(sequence, model, vocabulary = c("\\n", "a", "c",
  "g", "t"))
}
\arguments{
\item{sequence}{Input sequence, length should be in sync with the model.
If length exceeds input.shape of model then only the right side of the
sequence will be used.}

\item{model}{Trained model from the function \code{trainNetwork()}}

\item{vocabulary}{Ordered vocabulary of input sequence}
}
\description{
The output is a S4 class.
}
