% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictNextNucleotide}
\alias{predictNextNucleotide}
\title{Predict next nucleotides in sequence output as S4 class}
\usage{
predictNextNucleotide(sequence, model, vocabulary = c("\\n", "a", "c",
  "g", "t"))
}
\arguments{
\item{sequence}{input sequence, length should be in sync with the model
if length exceeds input.shape of model then only the right side of the
sequence will be used}

\item{model}{trained lstm}

\item{vocabulary}{ordered vocabulary of input sequence}
}
\description{
Predict next nucleotides in sequence output as S4 class
}
