% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessFasta}
\alias{preprocessFasta}
\title{wrapper of the preprocessSemiRedundant() function called on the genomic contents of one
fasta file. Multiple entries are combined with newline characters.}
\usage{
preprocessFasta(path, labels = NULL, maxlen = 250,
  vocabulary = c("\\n", "a", "c", "g", "t"), verbose = F)
}
\description{
wrapper of the preprocessSemiRedundant() function called on the genomic contents of one
fasta file. Multiple entries are combined with newline characters.
}
