% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessSemiRedundant}
\alias{preprocessSemiRedundant}
\title{Preprocess string to semi redundant one-hot vector}
\usage{
preprocessSemiRedundant(char, labels = NULL, maxlen = 250, vocabulary,
  verbose = F)
}
\arguments{
\item{char}{a character string of text, length of one}

\item{labels}{a character string of same length as char with character as labels}

\item{maxlen}{length of semi-redundant sequences of maxlen characters}

\item{vocabulary}{char, should be sorted, if not set char vocabulary will be used}

\item{verbose}{TRUE/FALSE}
}
\description{
Outputs semi-redundant set of input string
Preprocess a character input
Collapse and tokenize and vectorize character
Use this function with a character string as input such as data(train)
if the input text ist ABCDEFGHIJKLM and the maxlen is set to 5, the chunks would be
X(1): ABCDE Y(1):F
X(2): BCDEF Y(2):G
X(3): CDEFG Y(3):H
X(4): DEFGH (4):I
...
}
