% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{reduce_lr_cb}
\alias{reduce_lr_cb}
\title{Learning_rate callback}
\usage{
reduce_lr_cb(patience, cooldown, lr_plateau_factor, monitor = "val_acc")
}
\arguments{
\item{patience}{Number of epochs waiting for decrease in validation loss before reducing learning rate.}

\item{cooldown}{Number of epochs without changing learning rate.}

\item{lr_plateau_factor}{Factor of decreasing learning rate when plateau is reached.}
}
\description{
Learning_rate callback
}
\keyword{internal}
