% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{remove_checkpoints}
\alias{remove_checkpoints}
\title{Remove checkpoints}
\usage{
remove_checkpoints(
  cp_dir,
  metric = "acc",
  best_n = 1,
  ask_before_remove = TRUE
)
}
\arguments{
\item{cp_dir}{Directory containing checkpoints.}

\item{metric}{Either \code{"acc"}, \code{"loss"} or \code{"last_ep"}. Condition which checkpoints to keep.}

\item{best_n}{Number of checkpoints to keep.}

\item{ask_before_remove}{Whether to show files to keep before deleting rest.}
}
\description{
Remove all but n 'best' checkpoints, based on some condition. Condition can be
accuracy, loss or epoch number.
}
