% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{remove_checkpoints}
\alias{remove_checkpoints}
\title{Remove checkpoints}
\usage{
remove_checkpoints(
  cp_dir,
  metric = "acc",
  best_n = 1,
  ask_before_remove = TRUE
)
}
\arguments{
\item{cp_dir}{Directory containing checkpoints.}

\item{metric}{Either `"acc"`, `"loss"` or `"last_ep"`. Condition which checkpoint to keep.}

\item{best_n}{Number of checkpoints to keep.}

\item{ask_before_remove}{Whether to show files to keep before deleting rest.}
}
\description{
Remove all but one checkpoint, based on some condition.
}
