% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{replaceChar}
\alias{replaceChar}
\title{Replaces specific nucleotides in a sequence one by one}
\usage{
replaceChar(sequence, model, char = "X", vocabulary = c("\\n", "a",
  "c", "g", "t"))
}
\arguments{
\item{sequence}{input sequence}

\item{model}{trained lstm}

\item{char}{character that will be replaced}

\item{vocabulary}{ordered vocabulary of input sequence}
}
\description{
Replaces specific nucleotides in a sequence one by one
}
