% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{replaceChar}
\alias{replaceChar}
\title{Replaces specific nucleotides in a sequence}
\usage{
replaceChar(sequence, model, char = "X", vocabulary = c("\\n", "a",
  "c", "g", "t"))
}
\arguments{
\item{sequence}{Input sequence, length should be in sync with the model.
If length exceeds input.shape of model then only the right side of the
sequence will be used.}

\item{model}{Trained model from the function \code{trainNetwork()}}

\item{char}{Character in the sequence that will be replaced}

\item{vocabulary}{Ordered vocabulary of input sequence}
}
\description{
Replaces specific nucleotides in a sequence
}
