% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{reset_states_cb}
\alias{reset_states_cb}
\title{Reset states at start/end of validation and whenever file changes}
\usage{
reset_states_cb(
  path_file_log,
  path_file_logVal,
  num_files_old = 0,
  num_files_new = 0
)
}
\description{
Reset states at start/end of validation and whenever file changes
}
