% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{reset_states_cb}
\alias{reset_states_cb}
\title{Reset states callback}
\usage{
reset_states_cb(path_file_log, path_file_logVal)
}
\arguments{
\item{path_file_log}{Path to log of training files.}

\item{path_file_logVal}{Path to log of validation files.}
}
\description{
Reset states at start/end of validation and whenever file changes. Can be used for stateful LSTM.
}
