% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R, R/create_model_utils.R
\name{reshape_input}
\alias{reshape_input}
\title{Replace input layer}
\usage{
reshape_input(model, input_shape)

reshape_input(model, input_shape)
}
\arguments{
\item{model}{A keras model.}

\item{input_shape}{The new input shape vector (without batch size).}
}
\value{
A keras model with changed input shape of input model.
}
\description{
Replace first layer of model with new input layer of different shape. Only works for sequential models that
use CNN and LSTM layers.

Replace first layer of model with new input layer of different shape. Only works for sequential models that
use CNN and LSTM layers.
}
\examples{
model_1 <-  create_model_lstm_cnn(
  maxlen = 50,
  kernel_size = c(10, 10),
  filters = c(64, 128),
  pool_size = c(2, 2),
  layer_lstm = c(32),
  verbose = FALSE,
  layer_dense = c(64, 2))
model <- reshape_input(model_1, input_shape = c(120, 4))
model
\dontshow{if (reticulate::py_module_available("tensorflow")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
model_1 <-  create_model_lstm_cnn(
  maxlen = 50,
  kernel_size = c(10, 10),
  filters = c(64, 128),
  pool_size = c(2, 2),
  layer_lstm = c(32),
  verbose = FALSE,
  layer_dense = c(64, 2))
model <- reshape_input(model_1, input_shape = c(120, 4))
model
\dontshow{\}) # examplesIf}
}
