% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{seq_encoding_lm}
\alias{seq_encoding_lm}
\title{Encodes integer sequence for language model}
\usage{
seq_encoding_lm(
  sequence,
  maxlen,
  vocabulary,
  start_ind,
  ambiguous_nuc = "zero",
  nuc_dist = NULL,
  use_quality = FALSE,
  quality_vector = NULL,
  target_len = 1,
  use_coverage = FALSE,
  max_cov = NULL,
  cov_vector = NULL,
  n_gram = NULL,
  n_gram_stride = 1,
  output_format = "target_right"
)
}
\arguments{
\item{sequence}{Sequence of integers.}

\item{maxlen}{Length of one sample}

\item{vocabulary}{Set of characters to encode.}

\item{start_ind}{Start positions of samples in \code{sequence}.}

\item{ambiguous_nuc}{How to handle nucleotides outside vocabulary, either "zero", "discard" or "equal". If "zero", input gets encoded as zero vector;
if "equal" input is 1/length(vocabulary) x length(vocabulary). If "discard" samples containing nucleotides outside vocabulary get discarded.}

\item{nuc_dist}{Nucleotide distribution.}

\item{use_quality}{Use quality scores.}

\item{quality_vector}{Vector of quality probabilities.}
}
\description{
\code{seq_encoding_lm} Helper function for \code{\link{generator_fasta_lm}}. 
Encodes integer sequence to input/target list according to \code{output_format} argument.
}
