% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{sequenceToArray}
\alias{sequenceToArray}
\title{Helper function for fastaFileGenerator}
\usage{
sequenceToArray(sequence, maxlen, vocabulary = c("\\n", "a", "c", "g",
  "t"))
}
\arguments{
\item{sequence}{character sequence}

\item{maxlen}{length of one sample}
}
\description{
Helper function for fastaFileGenerator
}
