% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{sequenceToArray}
\alias{sequenceToArray}
\title{Helper function for fastaFileGenerator}
\usage{
sequenceToArray(sequence, maxlen, vocabulary = c("-", "|", "a", "c", "g",
  "t"))
}
\arguments{
\item{sequence}{character sequence}

\item{maxlen}{length of one sample

Returns one hot encoding for every sequence  
For example: sequence = "acatg", maxlen = 4, leads to
X = (0 1 0 0 0  
     0 0 1 0 0
     0 1 0 0 0
     0 0 0 0 1)
Y = (0 0 0 1 0)}
}
\description{
Helper function for fastaFileGenerator
}
