% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{split_fasta}
\alias{split_fasta}
\title{Split fasta file into smaller files.}
\usage{
split_fasta(
  path_input,
  target_folder,
  split_n = 500,
  shuffle_entries = TRUE,
  delete_input = FALSE
)
}
\arguments{
\item{path_input}{Fasta file to split into smaller files}

\item{split_n}{Maximum number of entries to use in smaller file.}

\item{shuffle_entries}{Whether to shuffle fasta entries before split.}

\item{delete_input}{Whether to delete the original file.}

\item{target_path}{Directory for output.}
}
\description{
Returns smaller files with same file name and "_x" (where x is an integer). For example,
assume we have input file called "abc.fasta" with 100 entries and split_n = 50. Function will
create two files called "abc_1.fasta" and "abc_2.fasta" in target_path.
}
