% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu-management.R
\name{startGPUSession}
\alias{startGPUSession}
\title{Set # of GPUs to be used in TensorFlow backend.}
\usage{
startGPUSession(gpus = "0")
}
\arguments{
\item{gpus}{GPU ids}
}
\description{
Set # of GPUs to be used in TensorFlow backend.
}
