% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu-management.R
\name{startGPUSession}
\alias{startGPUSession}
\title{Limit Tensorflow proccess to one or multiple GPUs.}
\usage{
startGPUSession(gpus = "0")
}
\arguments{
\item{gpus}{GPU ids}
}
\description{
Limit Tensorflow proccess to one or multiple GPUs.
}
