% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{tsneFromStates}
\alias{tsneFromStates}
\title{run the tSNE on a cell from the states.list, return a deepTsne S4 object which contains the output of the Rtsne::Rtsne and the sample ids}
\usage{
tsneFromStates(states.list, cell.number = 1, flatten = FALSE)
}
\arguments{
\item{states.list}{list of states generated by generateStatesromFolder}

\item{cell.number}{number of cell/neuron that should go into tSNE}

\item{flatten}{if true, all cells will be used}
}
\description{
run the tSNE on a cell from the states.list, return a deepTsne S4 object which contains the output of the Rtsne::Rtsne and the sample ids
}
