% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{validation_after_training_cb}
\alias{validation_after_training_cb}
\title{Validation after training callback}
\usage{
validation_after_training_cb(gen.val, validation_steps)
}
\arguments{
\item{gen.val}{Validation generator}

\item{validation_steps}{Number of validation steps.}
}
\description{
Do validation only once at end of training.
}
