% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{validation_after_training_cb}
\alias{validation_after_training_cb}
\title{Validation after training callback}
\usage{
validation_after_training_cb(gen.val, validation_steps)
}
\arguments{
\item{gen.val}{Validation generator}

\item{validation_steps}{Number of validation steps.}
}
\value{
Keras callback, apply validation only after training.
}
\description{
Do validation only once at end of training.
}
\examples{
maxlen <- 20
model <- create_model_lstm_cnn(layer_lstm = 8, maxlen = maxlen)
gen <- get_generator(train_type = 'dummy_gen', model = model, batch_size = 4, maxlen = maxlen)
vat <- validation_after_training_cb(gen.val = gen, validation_steps = 10)

}
