% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writedict}
\alias{writedict}
\title{generates dictionary for LSTMVis}
\usage{
writedict(dat, filename = "train.dict")
}
\arguments{
\item{dat}{character}

\item{dat}{filename where dict file is written to}
}
\description{
generates dictionary for LSTMVis
}
