% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writehdf5}
\alias{writehdf5}
\title{generates hdf5 file containing the character id for each time step}
\usage{
writehdf5(dat, filename = "train.hdf5")
}
\arguments{
\item{dat}{character}

\item{dat}{filename where hdf5 file is written to}
}
\description{
generates hdf5 file containing the character id for each time step
}
